% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{appendix_addin}
\alias{appendix_addin}
\title{Insert the standard DDI assessment report appendix as mardown}
\usage{
appendix_addin()
}
\description{
Insert the standard DDI assessment report appendix as mardown
}
