% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{methods}
\name{as.data.frame,Organism-method}
\alias{as.data.frame,Organism-method}
\title{from.json.Organism}
\usage{
\S4method{as.data.frame}{Organism}(object, row.names = NULL,
  optional = FALSE, ...)
}
\value{
Data frame
}
\description{
This function converts an Organism json object to Organism
}

