% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyp.R
\name{basic_cyp_inhibition_risk.list}
\alias{basic_cyp_inhibition_risk.list}
\title{Basic CYP inhibition risk evaluation for a list of perpetrator objects}
\usage{
\method{basic_cyp_inhibition_risk}{list}(perp, cyp_inh)
}
\arguments{
\item{perp}{Perpetrator objects as list.}

\item{cyp_inh}{CYP inhibition data as data frame.}
}
\value{
A list of data frames.
}
\description{
Basic CYP inhibition risk evaluation for a list of perpetrator objects
}
