% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyp.R
\name{basic_cyp_inhibition_risk.perpetrator}
\alias{basic_cyp_inhibition_risk.perpetrator}
\title{Basic CYP inhibition risk evaluation for a perpetrator object}
\usage{
\method{basic_cyp_inhibition_risk}{perpetrator}(perp, cyp_inh)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{cyp_inh}{CYP inhibiton data as data frame.}
}
\value{
A data frame.
}
\description{
Basic CYP inhibition risk evaluation for a perpetrator object
}
