% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyp.R
\name{basic_cyp_inhibition_risk_table.list}
\alias{basic_cyp_inhibition_risk_table.list}
\title{Table of basic CYP inhibition risk}
\usage{
\method{basic_cyp_inhibition_risk_table}{list}(perp, cyp_inh, na.rm = F)
}
\arguments{
\item{perp}{A list of perpetrator objects.}

\item{cyp_inh}{CYP inhibition data as data frame.}

\item{na.rm}{Remove rows with lacking ki data (i.e., where ki == NA).}
}
\value{
Basic evaluation of CYP inhibition risk as markdown-formatted tables,
  or an empty string if no CYP inhibition data available.
}
\description{
Table of basic CYP inhibition risk
}
