% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyp.R, R/risk.R
\name{basic_cyp_tdi_risk}
\alias{basic_cyp_tdi_risk}
\title{Basic modeling of the CYP time-dependent inhibition risk}
\usage{
basic_cyp_tdi_risk(perp, cyp_tdi, cyp_kdeg = cyp_turnover)

basic_cyp_tdi_risk(perp, cyp_tdi, cyp_kdeg = cyp_turnover)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{cyp_tdi}{The CYP TDI data as data frame. The following fields are expected:
\itemize{
\item 'name' The perpetrator compound name as character.
\item 'cyp' The CYP enzyme as character.
\item 'ki' The \eqn{K_I} in \eqn{\mu M} as numeric.
\item 'kinact' The \eqn{k_{inact}} in 1/h as numeric.
\item 'source' Optional source information as character,
}}

\item{cyp_kdeg}{}
}
\value{
A data frame.
}
\description{
This function calculates the risk for time-dependent inhibition of CYP
enzymes.
}
\details{
The risk assessment is based on:

\deqn{R=\frac {k_{obs} + k_{deg}}{k_{deg}}}

where

\deqn{k_{obs}=\frac {5*k_{inact}*C_{max,u}}{K_{I,u} + 5 * C_{max,u}}}

Values of \eqn{R > 1.25} suggest a relevant TDI potential.

The CYP degradation rates, \eqn{k_{deg}} are physiological constants that
should be derived from the scientific literature. This package provides
standard values for \eqn{k_{deg}} in \link{cyp_turnover} that are commonly used.

In the present version, only the risk for hepatic TDI for CYP enzymes is
calculated.
}
\examples{
basic_cyp_tdi_risk(examplinib_parent, examplinib_cyp_tdi_data)
}
\seealso{
\link{cyp_turnover}

\code{\link[=read_tdi_data]{read_tdi_data()}}
}
