% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{compound_names_string}
\alias{compound_names_string}
\title{Compound names as single string}
\usage{
compound_names_string(compounds)
}
\arguments{
\item{compounds}{A list of perpetrator objects}
}
\value{
A character string.
}
\description{
This function returns a single string with the comma-separated names of the
compounds included in the list of compunds.
}
\examples{
compound_names_string(examplinib_compounds)
}
