% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{conc_table}
\alias{conc_table}
\title{Key perpetrator concentrations}
\usage{
conc_table(perp)
}
\arguments{
\item{perp}{The object (compound object or list thereof)}
}
\description{
Key perpetrator concentrations
}
\examples{
conc_table(examplinib_parent)

}
\seealso{
\code{\link[=key_concentrations]{key_concentrations()}}
}
