% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perpetrator.R
\name{conctbl,compound-method}
\alias{conctbl,compound-method}
\title{Title}
\usage{
\S4method{conctbl}{compound}(x, round = 2)
}
\arguments{
\item{compound}{Compound object}
}
\value{
Markdown-formatted text
}
\description{
Title
}
