% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perpetrator.R
\name{conctbl}
\alias{conctbl}
\title{Title}
\usage{
conctbl(x, round = 2)
}
\arguments{
\item{x}{Compound object}

\item{round}{Markdown-formatted text}
}
\value{
Markdown-formatted text
}
\description{
Title
}
