% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{conditional_dose_string}
\alias{conditional_dose_string}
\title{Perpetrator dose as string}
\usage{
conditional_dose_string(perp, show_dose = TRUE)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{show_dose}{Show dose as logical.}
}
\value{
Character string.
}
\description{
Perpetrator dose as string
}
\examples{
conditional_dose_string(examplinib_parent, TRUE)
conditional_dose_string(examplinib_metabolite, TRUE)
conditional_dose_string(examplinib_parent, FALSE)
}
\keyword{internal}
