% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statements.R
\name{cyp_induction_risk_summary.list}
\alias{cyp_induction_risk_summary.list}
\title{Title}
\usage{
\method{cyp_induction_risk_summary}{list}(perp, ...)
}
\arguments{
\item{perp}{A list of perpetrator objects.}

\item{...}{Further arguments.}
}
\value{
Text output.
}
\description{
Title
}
