% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statements.R
\name{cyp_induction_risk_summary.perpetrator}
\alias{cyp_induction_risk_summary.perpetrator}
\title{Summary statement for CYP induction for a precipitant compound}
\usage{
\method{cyp_induction_risk_summary}{perpetrator}(perp, cyp_ind, d = 1)
}
\arguments{
\item{cyp_ind}{}

\item{d}{}
}
\value{
Markdown-formatted text output.
}
\description{
Summary statement for CYP induction for a precipitant compound
}
