% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statements.R
\name{cyp_inhibition_risk_summary.perpetrator}
\alias{cyp_inhibition_risk_summary.perpetrator}
\title{Summary statement for CYP inhibition for a precipitant compound}
\usage{
\method{cyp_inhibition_risk_summary}{perpetrator}(perp, cyp_inh)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{cyp_inh}{}
}
\value{
Markdown-formatted text output.
}
\description{
Summary statement for CYP inhibition for a precipitant compound
}
