% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{cyp_inhibitor_data}
\alias{cyp_inhibitor_data}
\title{Make CYP inhibition table}
\usage{
cyp_inhibitor_data(target, ki, source = character(0))
}
\arguments{
\item{target}{Name of the target.}

\item{ki}{Named list.}

\item{source}{Named list.}
}
\value{
A data table.
}
\description{
Make CYP inhibition table
}
\examples{
cyp_inhibitor_data(
"examplinib",
c(CYP3A4 = 1, CYP1A2 = 2, CYP2D6 = 3),
source = c(CYP3A4 = "source 1", CYP1A2 = "source 2"))

}
