% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statements.R
\name{cyp_tdi_risk_summary.list}
\alias{cyp_tdi_risk_summary.list}
\title{Summary statement for CYP TDI}
\usage{
\method{cyp_tdi_risk_summary}{list}(perp, ...)
}
\arguments{
\item{perp}{The perpetrator object.}
}
\value{
Text output.
}
\description{
Summary statement for CYP TDI
}
\examples{
cyp_tdi_risk_summary(examplinib_compounds, examplinib_cyp_tdi_data)
}
