% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statements.R
\name{cyp_tdi_risk_summary.perpetrator}
\alias{cyp_tdi_risk_summary.perpetrator}
\title{Summary statement for CYP TDI for a precipitant compound}
\usage{
\method{cyp_tdi_risk_summary}{perpetrator}(perp, cyp_tdi, cyp_kdeg = cyp_turnover)
}
\arguments{
\item{cyp_kdeg}{}
}
\value{
Markdown-formatted text output.
}
\description{
Summary statement for CYP TDI for a precipitant compound
}
