% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statements.R
\name{cyp_tdi_risk_summary}
\alias{cyp_tdi_risk_summary}
\title{Summary statement for CYP TDI}
\usage{
cyp_tdi_risk_summary(perp, cyp_tdi, cyp_kdeg = cyp_turnover)
}
\arguments{
\item{cyp_kdeg}{}
}
\value{
Markdown-formatted text output.
}
\description{
Summary statement for CYP TDI
}
\examples{
cyp_tdi_risk_summary(examplinib_parent, examplinib_cyp_tdi_data)
cyp_tdi_risk_summary(examplinib_compounds, examplinib_cyp_tdi_data)
}
