% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cyp_turnover}
\alias{cyp_turnover}
\title{CYP turnover rate constants}
\format{
A data frame with 3 columns
}
\source{
This data set is taken from various literature sources
}
\usage{
cyp_turnover
}
\description{
Mean degradation constants in 1/h for hepatic and intestinal CYP enzymes.
}
\details{
\preformatted{
    cyp kdeg_hepatic kdeg_intestinal
 CYP1A1       0.0183              NA
 CYP1A2       0.0183              NA
 CYP2A6       0.0267              NA
 CYP2B6       0.0217              NA
 CYP2C8       0.0301              NA
 CYP2C9       0.0067            0.03
CYP2C18       0.0267              NA
CYP2C19       0.0267            0.03
 CYP2D6       0.0099            0.03
 CYP2E1       0.0176              NA
 CYP2J2       0.0194            0.03
 CYP3A4       0.0193            0.03
 CYP3A5       0.0193            0.03
 CYP3A7       0.0193              NA
}
}
\seealso{
\code{\link[=basic_cyp_tdi_risk]{basic_cyp_tdi_risk()}}

\code{\link[=mech_stat_cyp_risk]{mech_stat_cyp_risk()}}
}
\keyword{datasets}
