% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{database,DatasetSummary-method}
\alias{database,DatasetSummary-method}
\title{Returns an acession id}
\usage{
\S4method{database}{DatasetSummary}(object)
}
\arguments{
\item{object}{a DatasetSummary}
}
\value{
database accession
}
\description{
Returns an acession id
}
\author{
Yasset Perez-Riverol
}
