% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{methods}
\name{dataset.id,DatasetDetail-method}
\alias{dataset.id,DatasetDetail-method}
\title{Returns an acession id}
\usage{
\S4method{dataset.id}{DatasetDetail}(object)
}
\arguments{
\item{object}{a DatasetDetail}
}
\value{
the accession
}
\description{
Returns an acession id
}
\author{
Yasset Perez-Riverol
}

