% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset.id,DatasetSummary-method}
\alias{dataset.id,DatasetSummary-method}
\title{Returns an acession id}
\usage{
\S4method{dataset.id}{DatasetSummary}(object)
}
\arguments{
\item{object}{a DatasetSummary}
}
\value{
the accession
}
\description{
Returns an acession id
}
\author{
Yasset Perez-Riverol
}
