% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericFunctions.R
\name{dataset.to.dataframe}
\alias{dataset.to.dataframe}
\title{Returns a data frame from a DatasetDetail list}
\usage{
dataset.to.dataframe(object, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{row.names}{optional row namesnt}

\item{optional}{optional}

\item{DatasetList}{instance}
}
\description{
Returns a data frame from a DatasetDetail list
}
