% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DatasetDetail-class}
\alias{DatasetDetail-class}
\title{DatasetDetail provides the information about an specific Facet in the API}
\description{
DatasetDetail provides the information about an specific Facet in the API
}
