% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DataSetResult-class}
\alias{DataSetResult-class}
\title{DataSetReult provides a list of datasets for an specific query}
\description{
DataSetReult provides a list of datasets for an specific query
}

