% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DatasetSimilars-class}
\alias{DatasetSimilars-class}
\title{DatasetSimilars, This class contains the information of one particuar dataset in OmicsDI}
\description{
DatasetSimilars, This class contains the information of one particuar dataset in OmicsDI
}

