% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DatasetSummary-class}
\alias{DatasetSummary-class}
\title{DatasetSummary, This class contains the information of one particuar dataset in OmicsDI}
\description{
DatasetSummary, This class contains the information of one particuar dataset in OmicsDI
}

