% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statements.R
\name{ddi_risk_summary}
\alias{ddi_risk_summary}
\title{Overall DDI risk summary text block}
\usage{
ddi_risk_summary(
  perp,
  cyp_inh = NULL,
  cyp_tdi = NULL,
  cyp_ind = NULL,
  ugt_inh = NULL,
  transporter_inh = NULL,
  d = 1,
  include_induction = TRUE,
  substr = cyp_reference_substrates,
  cyp_kdeg = cyp_turnover
)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{cyp_inh}{}

\item{cyp_tdi}{The CYP TDI data as data frame. The following fields are expected:
\itemize{
\item 'name' The perpetrator compound name as character.
\item 'cyp' The CYP enzyme as character.
\item 'ki' The \eqn{K_I} in \eqn{\mu M} as numeric.
\item 'kinact' The \eqn{k_{inact}} in 1/h as numeric.
\item 'source' Optional source information as character,
}}

\item{cyp_ind}{The CYP induction data as data frame. The following fields
are expected:
\itemize{
\item 'name' The name of the perpetrator compound as character.
\item 'cyp' The CYP enzyme as (upper case) character.
\item 'emax' The \eqn{E_{max}}, i.e., the maximum induction effect determined in
vitro as numeric.
\item 'ec50' The \eqn{EC_{50}} in \eqn{\mu M} as numeric.
\item 'maxc' The maximal concentration in \eqn{\mu M} tested in the in vitro assay as
numeric.
\item 'source' Optional source information as character.
}}

\item{ugt_inh}{}

\item{transporter_inh}{Transporter inhibition data as data frame. The
following fields are expected:
\itemize{
\item 'name' The perpetrator compound name.
\item 'cyp' The UGT enzyme as (upper case) character.
\item 'ic50' The \eqn{IC_{50}} of the inhibition in μM.
\item 'source' Optional source information as character.
}}

\item{d}{Scaling factor, defaults to 1.}

\item{include_induction}{Switch to define whether induction effects should be
included in the calculation (C-terms as per the FDA guideline)}

\item{substr}{The CYP probe substrates to be used as data frame, defaults to
\link{cyp_reference_substrates}. The data frame is expected to have the following
fields:
\itemize{
\item 'cyp' The CYP enzyme as (upper case) character.
\item 'substrate' The substrate name as character.
\item 'fgut' The fraction of the drug escaping gut metabolism.
\item 'fm' The fraction of the drug that undergoes hepatic metabolism.
\item 'fmcyp' The fraction metabolized by the respective CYP enzyme.
}}

\item{cyp_kdeg}{}
}
\value{
Markdown-formatted text output.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
ddi_risk_summary(examplinib_compounds, examplinib_cyp_inhibition_data,
  examplinib_cyp_tdi_data, examplinib_cyp_induction_data,
  examplinib_ugt_inhibition_data, examplinib_transporter_inhibition_data)
ddi_risk_summary(examplinib_parent, examplinib_cyp_inhibition_data)
}
