% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddir-package.R
\docType{package}
\name{ddir-package}
\alias{ddir}
\alias{ddir-package}
\title{ddir: Assess the Clinical Drug-Drug Interaction Risk for Drug Compounds}
\description{
The goal of ddir is to facilitate the analysis of the drug-drug interaction risk for drugs as per the relevant FDA and EMA guidelines.
}
\keyword{internal}
