% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{ddir_example}
\alias{ddir_example}
\title{Get example data path}
\usage{
ddir_example(path = NULL)
}
\arguments{
\item{path}{Subdirectory as character.}
}
\value{
Character.
}
\description{
Get example data path
}
