% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{df_to_string}
\alias{df_to_string}
\title{Render data frame to character.}
\usage{
df_to_string(df, indent = "", n = NULL, colnames = TRUE)
}
\arguments{
\item{df}{The datavframe.}

\item{indent}{A string that defines the left indentation of the rendered
output.}

\item{n}{The number of lines to be rendered. If NULL (default), all lines
are rendered.}

\item{colnames}{Boolean value to indicate whether column names are to be
included in the output.}
}
\value{
The data frame representation as character.
}
\description{
This function renders a data frame into a string object.
}
