% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{dose.list}
\alias{dose.list}
\title{Title}
\usage{
\method{dose}{list}(obj)
}
\arguments{
\item{obj}{List of perpetrator objects.}
}
\value{
Text output.
}
\description{
Title
}
\examples{
dose(examplinib_compounds)
}
