% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{dose}
\alias{dose}
\title{Oral dose}
\usage{
dose(obj)
}
\arguments{
\item{obj}{A perpetrator object or list thereof.}
}
\value{
Oral dose(s) as character
}
\description{
Oral dose
}
\examples{
dose(examplinib_parent)
dose(examplinib_metabolite)
dose(examplinib_compounds)
}
