% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/induction.R
\name{dplyr_reconstruct.induction_data}
\alias{dplyr_reconstruct.induction_data}
\title{Reconstruct inhibition_data after dplyr verbs}
\usage{
\method{dplyr_reconstruct}{induction_data}(data, template)
}
\arguments{
\item{data}{The data as data frame.}

\item{template}{The object template.}
}
\value{
A inhibition_data object.
}
\description{
Reconstruct inhibition_data after dplyr verbs
}
