% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{dplyr_reconstruct.risk}
\alias{dplyr_reconstruct.risk}
\title{reconstruct risk object}
\usage{
\method{dplyr_reconstruct}{risk}(data, template)
}
\arguments{
\item{data}{The risk table.}

\item{template}{The attributes.}
}
\value{
A risk object.
}
\description{
reconstruct risk object
}
