% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_compounds}
\alias{examplinib_compounds}
\title{Fictional substrate data}
\format{
## `examplinib_compounds`
A list of two: `perpetrator`objects:
* examplinib, a fictional drug.
* M1, a fictional metabolite of examplinib.
}
\source{
Fictional data, made up for demo purposes.
}
\usage{
examplinib_compounds
}
\description{
Fictional substrate data
}
\seealso{
[perpetrator()]

[load_perpetrators()]
}
\keyword{datasets}
