% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_compounds_string}
\alias{examplinib_compounds_string}
\title{Fictional substrate data as string}
\format{
An object of class \code{character} of length 1.
}
\source{
Fictional data, made up for demo purposes.
}
\usage{
examplinib_compounds_string
}
\description{
A text string containing data for two: `perpetrator` objects:
* examplinib, a fictional drug.
* M1, a fictional metabolite of examplinib.
}
\seealso{
[read_perpetrators()]
}
\keyword{datasets}
