% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_cyp_ind_metabolite}
\alias{examplinib_cyp_ind_metabolite}
\title{Examplinib metabolite CYP induction data}
\format{
An object of class \code{inducer} of length 1.
}
\source{
Fictional data, made up for demo purposes.
}
\usage{
examplinib_cyp_ind_metabolite
}
\description{
Examplinib metabolite CYP induction data
}
\keyword{datasets}
