% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_cyp_induction}
\alias{examplinib_cyp_induction}
\title{Examplinib CYP induction data}
\format{
An object of class \code{induction_data} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 7 rows and 5 columns.
}
\source{
Fictional data, made up for demo purposes.
}
\usage{
examplinib_cyp_induction
}
\description{
Examplinib CYP induction data
}
\keyword{datasets}
