% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_cyp_induction_data}
\alias{examplinib_cyp_induction_data}
\title{Examplinib CYP induction data}
\format{
An object of class \code{data.frame} with 14 rows and 6 columns.
}
\source{
Fictional data, made up for demo purposes.
}
\usage{
examplinib_cyp_induction_data
}
\description{
Examplinib CYP induction data
}
\details{
\preformatted{
      name     cyp  emax ec50 maxc    source
examplinib  CYP1A2  1.00   NA    5 study 007
examplinib  CYP2B6  1.00   NA    5 study 007
examplinib  CYP2C8    NA   NA   NA
examplinib  CYP2C9    NA   NA   NA
examplinib CYP2C19    NA   NA   NA
examplinib  CYP2D6    NA   NA   NA
examplinib  CYP3A4  7.35 1.64    3 study 007
        M1  CYP1A2  1.00   NA    5 study 007
        M1  CYP2B6  6.98 1.86    5 study 007
        M1  CYP2C8    NA   NA   NA
        M1  CYP2C9    NA   NA   NA
        M1 CYP2C19    NA   NA   NA
        M1  CYP2D6    NA   NA   NA
        M1  CYP3A4 22.70 1.10    5 study 007
}
}
\keyword{datasets}
