% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_cyp_inh_parent}
\alias{examplinib_cyp_inh_parent}
\title{Examplinib CYP inhibition data}
\format{
An object of class \code{inhibitor} of length 1.
}
\source{
Fictional data, made up for demo purposes.
}
\usage{
examplinib_cyp_inh_parent
}
\description{
Examplinib CYP inhibition data
}
\details{
CYP inhibition data for examplinib
}
\keyword{datasets}
