% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_cyp_inhibition}
\alias{examplinib_cyp_inhibition}
\title{Examplinib CYP inhibition data}
\format{
An object of class \code{inhibition_data} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 7 rows and 3 columns.
}
\source{
Fictional data, made up for demo purposes.
}
\usage{
examplinib_cyp_inhibition
}
\description{
Examplinib CYP inhibition data
}
\details{
CYP inhibition data for examplinib
}
\keyword{datasets}
