% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_cyp_inhibition_data}
\alias{examplinib_cyp_inhibition_data}
\title{Examplinib CYP inhibition data}
\format{
\subsection{\code{examplinib_cyp_inhibition_data}}{

A data frame with the columns \code{name}, \code{param}, \code{value} and \code{source}, where:
\itemize{
\item name is the name of the compound for which the data is recorded
\item param contains the respective CYP enzyme names.
\item value contains the Ki values for the respective CYP enzyme
\item source provides information for the source of the respective value, often
the name of the DMPK study. This entry is optional.
}
}
}
\source{
Fictional data, made up for demo purposes.
}
\usage{
examplinib_cyp_inhibition_data
}
\description{
This data frame is a typical input to the following functions:
\code{\link[=basic_cyp_inhibition_risk]{basic_cyp_inhibition_risk()}}
\link{basic_cyp_inhibition_risk_table}
\code{\link[=mech_stat_cyp_risk]{mech_stat_cyp_risk()}}
\code{\link[=mech_stat_cyp_risk_table]{mech_stat_cyp_risk_table()}}
}
\details{
CYP inhibition data can contain ki data for multiple compounds.
\preformatted{
      name   param      value    source
        M1    name         M1
        M1  CYP2C9        4.4 study 002
examplinib    name examplinib
examplinib  CYP1A2         NA
examplinib  CYP2B6         NA
examplinib  CYP2C8         11 study 001
examplinib  CYP2C9       13.5 study 001
examplinib CYP2C19         15 study 001
examplinib  CYP2D6         NA
examplinib  CYP3A4       12.5 study 001
}
}
\keyword{datasets}
