% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_cyp_inhibition_string}
\alias{examplinib_cyp_inhibition_string}
\title{Examplinib CYP inhibitor data as string}
\format{
An object of class \code{character} of length 1.
}
\source{
Fictional data for demo purposes.
}
\usage{
examplinib_cyp_inhibition_string
}
\description{
A character string containing CYP inhibition data for examplinib and M1.
}
\seealso{
\code{\link[=read_inhibitor_data]{read_inhibitor_data()}}
}
\keyword{datasets}
