% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_cyp_tdi_data}
\alias{examplinib_cyp_tdi_data}
\title{Examplinib CYP TDI data}
\format{
An object of class \code{data.frame} with 1 rows and 5 columns.
}
\source{
Fictional data, made up for demo purposes.
}
\usage{
examplinib_cyp_tdi_data
}
\description{
Examplinib CYP TDI data
}
\details{
\preformatted{
      name    cyp   ki kinact    source
examplinib CYP3A4 30.7   0.04 study 001
}
}
\keyword{datasets}
