% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_cyp_tdi_parent}
\alias{examplinib_cyp_tdi_parent}
\title{Examplinib CYP TDI data}
\format{
An object of class \code{inhibitor} of length 1.
}
\source{
Fictional data, made up for demo purposes.
}
\usage{
examplinib_cyp_tdi_parent
}
\description{
Examplinib CYP TDI data
}
\details{
\preformatted{
      name    cyp   ki kinact    source
examplinib CYP3A4 30.7   0.04 study 001
}
}
\keyword{datasets}
