% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_cyp_tdi_string}
\alias{examplinib_cyp_tdi_string}
\title{Examplinib CYP TDI data as string}
\format{
An object of class \code{character} of length 1.
}
\source{
Fictional data for demo purposes.
}
\usage{
examplinib_cyp_tdi_string
}
\description{
A character string containing CYP TDI data for examplinib.
}
\seealso{
\code{\link[=read_tdi_data]{read_tdi_data()}}
}
\keyword{datasets}
