% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_in_vitro_tdi}
\alias{examplinib_in_vitro_tdi}
\title{in vitro sample data for CYP3A4 TDI}
\format{
A data frame with 4 columns
\itemize{
\item TIME Pre-incubation time in minutes.
\item CONC Precipitant concentration in uM.
\item ACT Enzyme activity in percent of initial.
\item SOURCE Source study report
}
}
\source{
Fully synthetic data set to demonstrate TDI-related calculations
}
\usage{
examplinib_in_vitro_tdi
}
\description{
CYP3A4 activity data at different pre-incubation timees and
different concentrations of a time-dependent inhibitor.
}
\keyword{datasets}
