% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_parent}
\alias{examplinib_parent}
\title{Perpetrator compound data for examplinib}
\format{
An object of class \code{perpetrator} of length 1.
}
\source{
Fictional data, made up for demo purposes.
}
\usage{
examplinib_parent
}
\description{
Perpetrator compound data for examplinib
}
\details{
\preformatted{
                param      value        source
name             name examplinib
oral             oral       TRUE
mw                 mw      492.6
dose             dose        450 clinical dose
imaxss         imaxss       3530     study 001
fu                 fu      0.023     study 002
fumic           fumic          1       default
rb                 rb          1     study 003
fa                 fa       0.81     study 003
fg                 fg          1       default
ka                 ka    0.00267       unknown
solubility solubility        Inf       default
}
}
\keyword{datasets}
