% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_transporter_inh_parent}
\alias{examplinib_transporter_inh_parent}
\title{Examplinib transporter inhibition data}
\format{
An object of class \code{inhibitor} of length 1.
}
\source{
Fictional data, made up for demo purposes.
}
\usage{
examplinib_transporter_inh_parent
}
\description{
Examplinib transporter inhibition data
}
\keyword{datasets}
