% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_transporter_inhibition_data}
\alias{examplinib_transporter_inhibition_data}
\title{Fictional transporter inhibition data}
\format{
An object of class \code{data.frame} with 12 rows and 4 columns.
}
\usage{
examplinib_transporter_inhibition_data
}
\description{
Fictional transporter inhibition data
}
\keyword{datasets}
