% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_ugt_inhibition}
\alias{examplinib_ugt_inhibition}
\title{Examplinib UGT inhibition data}
\format{
An object of class \code{inhibition_data} (inherits from \code{data.frame}) with 8 rows and 3 columns.
}
\source{
Fictional data, made up for demo purposes.
}
\usage{
examplinib_ugt_inhibition
}
\description{
Examplinib UGT inhibition data
}
\keyword{datasets}
