% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_ugt_inhibition_data}
\alias{examplinib_ugt_inhibition_data}
\title{Fictional UGT inhibition data}
\format{
An object of class \code{data.frame} with 18 rows and 4 columns.
}
\usage{
examplinib_ugt_inhibition_data
}
\description{
Fictional UGT inhibition data
}
\keyword{datasets}
