% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_ugt_inhibition_string}
\alias{examplinib_ugt_inhibition_string}
\title{Examplinib UGT inhibition data as string}
\format{
An object of class \code{character} of length 1.
}
\source{
Fictional data for demo purposes.
}
\usage{
examplinib_ugt_inhibition_string
}
\description{
A character string containing UGT inhibition data for examplinib and M1.
}
\seealso{
\code{\link[=read_ugt_inhibitor_data]{read_ugt_inhibitor_data()}}
}
\keyword{datasets}
