% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Facet-class}
\alias{Facet-class}
\title{Facet provides the information about an specific Facet in the API}
\description{
Facet provides the information about an specific Facet in the API
}

