% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{FacetValue-class}
\alias{FacetValue-class}
\title{FacetValue provides the information about an specific Facet in the API}
\description{
FacetValue provides the information about an specific Facet in the API
}

