% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{from.json.BiologicalSimilars}
\alias{from.json.BiologicalSimilars}
\title{from.json.BiologicalSimilars}
\usage{
from.json.BiologicalSimilars(json.object)
}
\arguments{
\item{similars}{json object}
}
\value{
Return the Biological Similars of a dataset
}
\description{
This function get information from DatasetResults
}

