% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{from.json.BiologicalSimilarsList}
\alias{from.json.BiologicalSimilarsList}
\title{from.json.BiologicalSimilarsList}
\usage{
from.json.BiologicalSimilarsList(accession = accession, database = database,
  similars = json.datasetDetail)
}
\arguments{
\item{accession}{accesioon of the current dataset}

\item{database}{Database}

\item{similars}{json object}
}
\value{
BiologicalSimilars List
}
\description{
This function get information from DatasetResults
}

