% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset.R
\name{from.json.DatasetDetail}
\alias{from.json.DatasetDetail}
\title{from.json.DatasetDetail This function converts a json DatasetDetail to a DatasetDetail}
\usage{
from.json.DatasetDetail(json.object)
}
\arguments{
\item{json.object}{a DatasetDetail json object}
}
\value{
DatasetDetail
}
\description{
from.json.DatasetDetail This function converts a json DatasetDetail to a DatasetDetail
}

