% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{from.json.DataSetResult}
\alias{from.json.DataSetResult}
\title{Returns a DatasetResult instance from a JSON string representation}
\usage{
from.json.DataSetResult(json.object)
}
\arguments{
\item{json_str}{The JSON object}

\item{file}{the name of a file to read the json_str from; this can also be a URL. Only one of json_str or file must be supplied.}

\item{method}{use the C implementation, or the older slower (and one day to be depricated) R implementation}

\item{unexpected.escape}{changed handling of unexpected escaped characters. Handling value should be one of "error", "skip", or "keep"; on unexpected characters issue an error, skip the character, or keep the character}
}
\value{
The DatasetResult instance
}
\description{
Returns a DatasetResult instance from a JSON string representation
}
\details{
TOD
}
\author{
Yasset Perez-Riverol
}

