% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset.R
\name{from.json.DatasetResults}
\alias{from.json.DatasetResults}
\title{from.json.DatasetResults}
\usage{
from.json.DatasetResults(json.object)
}
\arguments{
\item{json.object}{DatasetResult json object}
}
\value{
DatasetResult
}
\description{
This function get information from DatasetResults
}

