% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset.R
\name{from.json.DatasetSummary}
\alias{from.json.DatasetSummary}
\title{from.json.DatasetSummary}
\usage{
from.json.DatasetSummary(json.object)
}
\arguments{
\item{json.object}{}
}
\value{
DatasetSummary
}
\description{
This function converts a json object to a DatasetSummary
}

