% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset.R
\name{from.json.Facet}
\alias{from.json.Facet}
\title{from.json.Facet}
\usage{
from.json.Facet(json.object)
}
\arguments{
\item{json.object}{Facet json object}
}
\value{
Facet
}
\description{
This function convert a from.json.Facet to a Facet
}

