% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset.R
\name{from.json.FacetValue}
\alias{from.json.FacetValue}
\title{from.json.FacetValue}
\usage{
from.json.FacetValue(json.object)
}
\arguments{
\item{json.object}{This object is a facetValue json object}
}
\value{
FacetValue a Final class with the object
}
\description{
This function converts a json FacetValue object to FaceTvalue
}

