% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{from.json.Organism}
\alias{from.json.Organism}
\title{from.json.Organism}
\usage{
from.json.Organism(json.object)
}
\arguments{
\item{json.object}{input json object}
}
\value{
Organism
}
\description{
This function converts an Organism json object to Organism
}
