% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{from.json.Protocol}
\alias{from.json.Protocol}
\title{from.json.Protocol This function converts a json Protocol object to a Protocol object}
\usage{
from.json.Protocol(json.object)
}
\arguments{
\item{json.object}{This json object contain the information of a protocol}
}
\value{
Protocol Object
}
\description{
from.json.Protocol This function converts a json Protocol object to a Protocol object
}

