% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get.BiologicalSimilars}
\alias{get.BiologicalSimilars}
\title{get.BiologicalSimilars
Returns a Biological Molecules similarity for a Dataset}
\usage{
get.BiologicalSimilars(accession, database)
}
\arguments{
\item{accession}{id in the original repository}

\item{database}{the original database where this dataset has been store}
}
\value{
List of similars
}
\description{
get.BiologicalSimilars
Returns a Biological Molecules similarity for a Dataset
}
\author{
Yasset Perez-Riverol
}

