% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.dataset.link,DatasetDetail-method}
\alias{get.dataset.link,DatasetDetail-method}
\title{Returns the dataset full link}
\usage{
\S4method{get.dataset.link}{DatasetDetail}(x)
}
\arguments{
\item{object}{a DatasetDetail}
}
\value{
the dataset link
}
\description{
Returns the dataset full link
}
\author{
Enrique Audain
}
