% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.dataset.name,DatasetDetail-method}
\alias{get.dataset.name,DatasetDetail-method}
\title{Returns the original dataset title}
\usage{
\S4method{get.dataset.name}{DatasetDetail}(x)
}
\arguments{
\item{object}{a DatasetDetail}
}
\value{
the dataset title
}
\description{
Returns the original dataset title
}
\author{
Enrique Audain
}
