% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.dataset.omics,DatasetDetail-method}
\alias{get.dataset.omics,DatasetDetail-method}
\title{Returns the type of omics experiment (e.g. Proteomics, Genomics)}
\usage{
\S4method{get.dataset.omics}{DatasetDetail}(x)
}
\arguments{
\item{object}{a DatasetDetail}
}
\value{
the omics
}
\description{
Returns the type of omics experiment (e.g. Proteomics, Genomics)
}
\author{
Enrique Audain
}
