% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.dataset.tissues,DatasetDetail-method}
\alias{get.dataset.tissues,DatasetDetail-method}
\title{Returns tissues information described in the dataset}
\usage{
\S4method{get.dataset.tissues}{DatasetDetail}(x)
}
\arguments{
\item{object}{a DatasetDetail}
}
\value{
list of the tissues information
}
\description{
Returns tissues information described in the dataset
}
\author{
Enrique Audain
}
