% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset.R
\name{get.DatasetDetail}
\alias{get.DatasetDetail}
\title{Returns a DatasetDetail from OmicsDI}
\usage{
get.DatasetDetail(accession, database)
}
\arguments{
\item{accession}{id in the original repository}

\item{database}{the original database where this dataset has been store}
}
\value{
DatasetDetail
}
\description{
Returns a DatasetDetail from OmicsDI
}
\details{
TODO
}
\author{
Jose A. Dianes
}

