% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.description,DatasetDetail-method}
\alias{get.description,DatasetDetail-method}
\title{Returns the dataset description}
\usage{
\S4method{get.description}{DatasetDetail}(x)
}
\arguments{
\item{object}{a DatasetDetail}
}
\value{
the dataset description
}
\description{
Returns the dataset description
}
\author{
Enrique Audain
}
