% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.diseases,DatasetDetail-method}
\alias{get.diseases,DatasetDetail-method}
\title{Returns the diseases under study}
\usage{
\S4method{get.diseases}{DatasetDetail}(x)
}
\arguments{
\item{object}{a DatasetDetail}
}
\value{
the diseases under study
}
\description{
Returns the diseases under study
}
\author{
Enrique Audain
}
