% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.experiments.type,DatasetDetail-method}
\alias{get.experiments.type,DatasetDetail-method}
\title{Returns the experiments type described in the study}
\usage{
\S4method{get.experiments.type}{DatasetDetail}(x)
}
\arguments{
\item{object}{a DatasetDetail}
}
\value{
list of experiments type
}
\description{
Returns the experiments type described in the study
}
\author{
Enrique Audain
}
