% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.instruments,DatasetDetail-method}
\alias{get.instruments,DatasetDetail-method}
\title{Returns the instruments used to generate the dataset}
\usage{
\S4method{get.instruments}{DatasetDetail}(x)
}
\arguments{
\item{object}{a DatasetDetail}
}
\value{
the instruments
}
\description{
Returns the instruments used to generate the dataset
}
\author{
Enrique Audain
}
