% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.keywords,DatasetDetail-method}
\alias{get.keywords,DatasetDetail-method}
\title{Returns the keywords described in the study}
\usage{
\S4method{get.keywords}{DatasetDetail}(x)
}
\arguments{
\item{object}{a DatasetDetail}
}
\value{
list of study keywords
}
\description{
Returns the keywords described in the study
}
\author{
Enrique Audain
}
