% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset.R
\name{get.MetadataSimilars}
\alias{get.MetadataSimilars}
\title{get.MetadataSimilars}
\usage{
get.MetadataSimilars(database = "", accession = "")
}
\arguments{
\item{database}{}

\item{accession}{}
}
\value{
DatasetResults
}
\description{
Dataset Similars by Metadata
}

