% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.organisms,DatasetDetail-method}
\alias{get.organisms,DatasetDetail-method}
\title{Returns the organisms described in the dataset}
\usage{
\S4method{get.organisms}{DatasetDetail}(x)
}
\arguments{
\item{object}{a DatasetDetail}
}
\value{
Named list (accession:organism) of organisms described in the study
}
\description{
Returns the organisms described in the dataset
}
\author{
Enrique Audain
}
