% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.publication.date,DatasetDetail-method}
\alias{get.publication.date,DatasetDetail-method}
\title{Returns the dataset publication date}
\usage{
\S4method{get.publication.date}{DatasetDetail}(x)
}
\arguments{
\item{object}{a DatasetDetail}
}
\value{
the dataset publication date
}
\description{
Returns the dataset publication date
}
\author{
Enrique Audain
}
