% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.publication.ids,DatasetDetail-method}
\alias{get.publication.ids,DatasetDetail-method}
\title{Returns the publications IDs associated to the dataset}
\usage{
\S4method{get.publication.ids}{DatasetDetail}(x)
}
\arguments{
\item{object}{a DatasetDetail}
}
\value{
list of associated publication IDs
}
\description{
Returns the publications IDs associated to the dataset
}
\author{
Enrique Audain
}
