% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perpetrator.R
\name{igut}
\alias{igut}
\title{Maximal gut concentration}
\usage{
igut(x, molar = FALSE)
}
\arguments{
\item{x}{A perpetrator compound object.}

\item{molar}{Output in molar concentration (uM). Defaults to FALSE (ng/nl).}
}
\value{
Numeric.
}
\description{
Maximal gut concentration
}
