% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_tools.R
\name{indmod}
\alias{indmod}
\title{Analyze in vitro CYP induction data}
\usage{
indmod(x, precipitant = "")
}
\arguments{
\item{x}{In vitro induction data as data frame with the columns:
\itemize{
\item DONOR The hepatocyte donor
\item SAMPLE "test" or "positive_control"
\item CONC The preipitant concentration in uM.
\item OBJECT The DDI object.
\item FOLD The mRNA or activity fold change.
\item REL The fold change relative to positive control.
\item SOURCE Source information
}}

\item{precipitant}{The precipitant as character.}
}
\value{
A list.
}
\description{
Analyze in vitro CYP induction data
}
\examples{
indmod(examplinib_in_vitro_ind, precipitant = "examplinib")
}
