% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inducer.R
\name{inducer}
\alias{inducer}
\title{DDI CYP inducer object constructor}
\usage{
inducer(data, precipitant = "")
}
\arguments{
\item{data}{Data frame with the columns:
\itemize{
\item object: The CYP induction target. Must be one of "CYP1A1", "CYP1A2",
"CYP2A6", "CYP2B6", "CYP2C8", "CYP2C9", "CYP2C18", "CYP2C19", "CYP2D6",
"CYP2E1", "CYP2J2", "CYP3A4", "CYP3A5", or "CYP3A7".
\item emax: The maximal induction effect as numeric.
\item ec50: The EC50, i.e., the concentration causing the half-maximal induction
effect, in uM.
\item max_c: The maximal concentration tested in the assay in uM.
\item source: Source information for the parameter.
}}

\item{precipitant}{The name of the DDI perpetrator as character.}
}
\value{
Inducer object
}
\description{
DDI CYP inducer object constructor
}
\examples{
inducer(data.frame(object = "CYP3A4", emax = 2, ec50 = .1, max_c = 100,
source = "test source"), precipitant = "test")

inducer(data = NULL, precipitant = "test")
}
