% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/induction.R
\name{induction_data}
\alias{induction_data}
\title{induction_data constructor alias}
\usage{
induction_data(data = NULL, precipitant = "")
}
\arguments{
\item{data}{A data frame with the columns object, emax, ec50, max_c and source.}

\item{precipitant}{The precipitant name as character.}
}
\value{
An induction_data object.
}
\description{
\itemize{
\item emax: Fitted Emax
\item ec50: Fitted EC50 in uM
\item max_c: Maximally tested perpetrator concentration in uM.
}
}
