% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_vitro_inducer.R
\docType{class}
\name{induction_experiment-class}
\alias{induction_experiment-class}
\title{In vitro induction data set}
\description{
In vitro induction data set
}
\section{Slots}{

\describe{
\item{\code{precipitant}}{character.}

\item{\code{data}}{data.frame.}
}}

