% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_vitro_inducer.R
\name{induction_experiment}
\alias{induction_experiment}
\title{induction experiment data set constructor}
\usage{
induction_experiment(data = NULL, precipitant = "")
}
\arguments{
\item{data}{A data frame.}

\item{precipitant}{The precipitant.}
}
\value{
induction_experiment object.
}
\description{
induction experiment data set constructor
}
\examples{
induction_experiment(examplinib_in_vitro_ind, "examplinib")
}
