% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_tools.R
\name{induction_plot}
\alias{induction_plot}
\title{Plot in vitro CYP induction data}
\usage{
induction_plot(x, precipitant = "test", type = "REL")
}
\arguments{
\item{x}{The in vitro data set as data frame. The following fields are
expected:
\itemize{
\item DONOR The hepatocyte donor batch as character.
\item PRECIPITANT The DDI precipitant as character.
\item CONC The precipitant concentration.
\item OBJECT The DDI target object as character.
\item FOLD The mRNA or enzyme activity change at the respective precipitant
concentration.
\item REL The mRNA or enzyme activity induction, as percent of positive control
at the respective precipitant concentration.
}

FOLD and REL are optional but a column that corresponds to the 'type'
argument must be present.}

\item{precipitant}{The DDI precipitant, defaults to 'test'.}

\item{type}{The DDI metric, e.g. FOLD or REL, as character. A different type
can be given but a column of that name must be present in the data set.}
}
\value{
A ggplot object.
}
\description{
Plot in vitro CYP induction data
}
\examples{
induction_plot(examplinib_in_vitro_ind)
induction_plot(examplinib_in_vitro_ind, type = "FOLD")
}
