% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{is_igut_solubility_limited}
\alias{is_igut_solubility_limited}
\title{Test if Igut of a perpetrator is limited by its solubility}
\usage{
is_igut_solubility_limited(obj)
}
\arguments{
\item{obj}{The perpetrator object.}
}
\value{
A boolean value.
}
\description{
Test if Igut of a perpetrator is limited by its solubility
}
