% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{is_oral}
\alias{is_oral}
\title{Test whether perpetrator is subject to first-pass elimination}
\usage{
is_oral(obj)
}
\arguments{
\item{obj}{The perpetrator object.}
}
\value{
Logical
}
\description{
Test whether perpetrator is subject to first-pass elimination
}
\examples{
is_oral(examplinib_parent)
is_oral(examplinib_metabolite)
}
\keyword{internal}
