% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyp.R
\name{kinetic_cyp_induction_risk.perpetrator}
\alias{kinetic_cyp_induction_risk.perpetrator}
\title{Basic kinetic CYP induction risk}
\usage{
\method{kinetic_cyp_induction_risk}{perpetrator}(perp, cyp_ind)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{cyp_ind}{The CYP induction data as data frame.}
}
\value{
A data frame.
}
\description{
Basic kinetic CYP induction risk
}
