% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyp.R
\name{kinetic_cyp_induction_risk_table}
\alias{kinetic_cyp_induction_risk_table}
\title{#' Basic kinetic CYP induction risk
#'
#' @details
#' Basic kinetic modeling of the CYP induction risk considers:
#'
#' \deqn{R = \frac {1}{1 + d * \frac {E_{max}*10*C_{max,ss,u}}{EC_{50,u} + 10 * C_{max,ss,u}}}}
#'
#' \eqn{d} is a scaling factor with a standard value of 1. A different value can
#' be used if warranted by prior experience with the experimental conditions.
#'
#' \eqn{R \le 0.8} suggest a relevant in vivo CYP induction potential.
#'
#' @inheritParams static_cyp_induction_risk
#' @param d Scaling factor, defaults to 1.
#' @return A data frame.
#' @export
#' @examples
#' kinetic_cyp_induction_risk(examplinib_parent, examplinib_cyp_induction_data)
kinetic_cyp_induction_risk <- function(perp, cyp_ind, d = 1) {
i <- key_concentrations(perp, molar=TRUE)}
\usage{
kinetic_cyp_induction_risk_table(perp, cyp_ind, na.rm = F, show_dose = FALSE)
}
\arguments{
\item{na.rm}{Switch to define whether rows with lacking \eqn{E_{max}} data
are removed from the output. Defaults to \code{FALSE}.}

\item{show_dose}{Show perpetrator dose in table title, defaults to \code{FALSE.}}
}
\value{
A markdown-formatted table.
}
\description{

}
\examples{
kinetic_cyp_induction_risk_table(examplinib_parent,
  examplinib_cyp_induction_data)
kinetic_cyp_induction_risk_table(examplinib_compounds,
  examplinib_cyp_induction_data)
kinetic_cyp_induction_risk_table(examplinib_compounds,
  examplinib_cyp_induction_data, show_dose = TRUE)
}
\seealso{
\code{\link[=kinetic_cyp_induction_risk]{kinetic_cyp_induction_risk()}}
}
