% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{LabMember-class}
\alias{LabMember-class}
\title{LabMember retrieve the information of a lab member including affilation, name, etc}
\description{
LabMember retrieve the information of a lab member including affilation, name, etc
}

