% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{load_cyp_inducer_data}
\alias{load_cyp_inducer_data}
\title{Load CYP inducer data from file}
\usage{
load_cyp_inducer_data(filename)
}
\arguments{
\item{filename}{The filename.}
}
\value{
A list of data frames.
}
\description{
Load CYP inducer data from file
}
