% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{load_dmpk_data}
\alias{load_dmpk_data}
\title{Load DMPK data (CYP inhibition, transporter inhibition, etc.)}
\usage{
load_dmpk_data(filename)
}
\arguments{
\item{filename}{The filename.}
}
\value{
A list of data frames.
}
\description{
Load DMPK data (CYP inhibition, transporter inhibition, etc.)
}
