% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{load_perpetrators}
\alias{load_perpetrators}
\title{Load perpetrator file}
\usage{
load_perpetrators(filename)
}
\arguments{
\item{filename}{The full path to the compound file}
}
\value{
A list of perpetrator objects
}
\description{
Load perpetrator file
}
