% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statements.R
\name{mech_stat_cyp_risk_summary.list}
\alias{mech_stat_cyp_risk_summary.list}
\title{Summary statement for mechanistic-static modeling of CYP modulation}
\usage{
\method{mech_stat_cyp_risk_summary}{list}(perp, ...)
}
\arguments{
\item{perp}{A perpetrator object.}

\item{...}{Further parameters.}
}
\value{
Markdown-formatted text output.
}
\description{
Summary statement for mechanistic-static modeling of CYP modulation
}
