% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkDDI_metadata.R
\name{mkDDI_metadata}
\alias{mkDDI_metadata}
\title{Make a DDI meta data file from reserch informations}
\usage{
mkDDI_metadata(title, data.file.name, ddi.file.name, var.ID, var.name,
  var.qstnLit, var.catgry.catValu, var.catgry.labl)
}
\arguments{
\item{title}{single text object of the research title}

\item{data.file.name}{single text object of a data file name (including path information)}

\item{ddi.file.name}{single text object of a ddi file name (including path information)}

\item{var.ID}{list of variable IDs}

\item{var.name}{list of variable names}

\item{var.qstnLit}{list of questions}

\item{var.catgry.catValu}{list of categorical values (NA for non-categorical variables)}

\item{var.catgry.labl}{list of labels for categorical values (NA for non-categorical variables)}
}
\description{
Make a DDI meta data file from reserch informations
}
