% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{name.perpetrator}
\alias{name.perpetrator}
\title{Name of a perpetrator compound}
\usage{
\method{name}{perpetrator}(obj)
}
\arguments{
\item{obj}{The perpetrator object.}
}
\value{
The name of the perpetrator as character.
}
\description{
Name of a perpetrator compound
}
\examples{
name(examplinib_parent)
}
