% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{name}
\alias{name}
\title{Generic name method}
\usage{
name(obj)
}
\arguments{
\item{obj}{The perpetrator object.}
}
\value{
The name of the perpetrator as character.
}
\description{
Generic name method
}
