% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Organism-class}
\alias{Organism-class}
\title{Organism return an organism entity including its name and accession}
\description{
Organism return an organism entity including its name and accession
}

