% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perpetrator.R
\docType{class}
\name{perpetrator-class}
\alias{perpetrator-class}
\title{Perpetrator class definition}
\value{
A perpetrator class object.
}
\description{
Perpetrator class definition
}
\section{Slots}{

\describe{
\item{\code{name}}{character.}

\item{\code{oral}}{logical.}

\item{\code{mw}}{numeric.}

\item{\code{dose}}{numeric.}

\item{\code{imaxss}}{numeric.}

\item{\code{fu}}{numeric.}

\item{\code{fumic}}{numeric.}

\item{\code{rb}}{numeric.}

\item{\code{fa}}{numeric.}

\item{\code{fg}}{numeric.}

\item{\code{ka}}{numeric.}

\item{\code{solubility}}{numeric.}

\item{\code{source}}{character}
}}

