% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_vitro_inducer.R
\name{plot,induction_experiment-method}
\alias{plot,induction_experiment-method}
\title{Plot method for induction_experiment objects}
\usage{
\S4method{plot}{induction_experiment}(x, type = "FOLD", ...)
}
\arguments{
\item{x}{An \code{induction_experiment} object.}

\item{type}{Metric to plot, \code{"FOLD"} or \code{"REL"}.}

\item{...}{Further arguments.}
}
\value{
A ggplot object.
}
\description{
Plot method for induction_experiment objects
}
