% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_vitro_inducer.R
\name{plot.induction_experiment}
\alias{plot.induction_experiment}
\title{Plot generic for induction_experiment}
\usage{
\method{plot}{induction_experiment}(x, type = "FOLD", ...)
}
\arguments{
\item{x}{The induction_experiment object.}

\item{type}{Columns to plot on y axis.}

\item{...}{Further parameters.}
}
\description{
Plot generic for induction_experiment
}
