% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipitant.R
\name{precipitant}
\alias{precipitant}
\title{Precipitant class definition}
\usage{
precipitant(
  name = "",
  oral = TRUE,
  mw = 0,
  dose = 0,
  imaxss = 0,
  fu = 1,
  fumic = 1,
  rb = 1,
  fa = 1,
  fg = 1,
  ka = 0.1,
  solubility = Inf,
  source = character(0)
)
}
\arguments{
\item{name}{character.}

\item{oral}{logical.}

\item{mw}{numeric.}

\item{dose}{numeric.}

\item{imaxss}{numeric.}

\item{fu}{numeric.}

\item{fumic}{numeric.}

\item{rb}{numeric.}

\item{fa}{numeric.}

\item{fg}{numeric.}

\item{ka}{numeric.}

\item{solubility}{numeric.}

\item{source}{character}
}
\value{
A precipitant class object.
}
\description{
Precipitant class definition
}
\examples{
precipitant(
  name = "examplinib",
  oral = TRUE,
  mw = 492.6,
  dose = 450,
  imaxss = 3530,
  fu = 0.023,
  fumic = 1,
  rb = 1,
  fa = 0.81,
  fg = 1,
  ka = 0.00267,
  solubility = Inf,
  source = c(dose = "clinical dose", imaxss = "study 001", fu = "study 002")
)
}
