% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perpetrator.R
\name{print,compound-method}
\alias{print,compound-method}
\title{Title}
\usage{
\S4method{print}{compound}(x)
}
\arguments{
\item{compound}{Compound object}
}
\value{
markdown-formatted text
}
\description{
Title
}
