% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inducer.R
\name{print,inducer-method}
\alias{print,inducer-method}
\title{Print DDI inducer object}
\usage{
\S4method{print}{inducer}(x)
}
\arguments{
\item{inducer}{DDI inducer object}
}
\value{
Markdown-formatted text
}
\description{
Print DDI inducer object
}
