% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_vitro_inducer.R
\name{print,induction_experiment-method}
\alias{print,induction_experiment-method}
\title{Pring method for induction_experiment objects}
\usage{
\S4method{print}{induction_experiment}(x)
}
\arguments{
\item{x}{An \code{induction_experiment} object.}
}
\value{
Nothing.
}
\description{
Pring method for induction_experiment objects
}
