% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/induction.R
\name{print.induction_data}
\alias{print.induction_data}
\title{Print method for induction_data objects}
\usage{
\method{print}{induction_data}(x, ...)
}
\arguments{
\item{x}{The inhibition}

\item{...}{Further arguments.}
}
\value{
Nothing.
}
\description{
Print method for induction_data objects
}
