% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_vitro_inducer.R
\name{print.induction_experiment}
\alias{print.induction_experiment}
\title{Print generic for induction_experiment}
\usage{
\method{print}{induction_experiment}(x, ...)
}
\arguments{
\item{x}{The induction_experiment object.}

\item{...}{Further parameters.}
}
\description{
Print generic for induction_experiment
}
