% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perpetrator.R
\name{print.perpetrator}
\alias{print.perpetrator}
\title{Print perpetrator object as markdown}
\usage{
print.perpetrator(x, ...)
}
\arguments{
\item{x}{The perpetrator object.}

\item{...}{Further arguments.}
}
\value{
A markdown-formatted table.
}
\description{
Print perpetrator object as markdown
}
