% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipitant.R
\name{print.precipitant}
\alias{print.precipitant}
\title{Print precipitant object as markdown}
\usage{
\method{print}{precipitant}(x, ...)
}
\arguments{
\item{x}{The precipitant object.}

\item{...}{Further arguments.}
}
\value{
A markdown-formatted table.
}
\description{
Print precipitant object as markdown
}
