% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{print.risk}
\alias{print.risk}
\title{Print risk object}
\usage{
\method{print}{risk}(x, ..., signif = 3)
}
\arguments{
\item{x}{The risk object.}

\item{...}{Further arguments.}

\item{signif}{Number of significant digits.}
}
\value{
Nothing or a markdodwn-formatted table
}
\description{
Print risk object
}
