% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{property_table.list}
\alias{property_table.list}
\title{Compound properties for a list of compounds}
\usage{
\method{property_table}{list}(obj)
}
\arguments{
\item{obj}{A list of perpetrator objects.}
}
\description{
Compound properties for a list of compounds
}
\examples{
property_table(examplinib_compounds)
}
