% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{property_table.perpetrator}
\alias{property_table.perpetrator}
\title{Table of perpetrator properties}
\usage{
\method{property_table}{perpetrator}(obj)
}
\arguments{
\item{obj}{The perpetrator object.}
}
\value{
Perpetrator properties as markdown-formatted table.
}
\description{
Table of perpetrator properties
}
\examples{
property_table(examplinib_parent)

}
