% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{property_table}
\alias{property_table}
\title{Table of perpetrator drug properties}
\usage{
property_table(obj)
}
\arguments{
\item{obj}{The perpetrator object or a list of perpetrator objects.}
}
\value{
A markdown-formatted table.
}
\description{
This function generates a markdown-formatted table of the key properties of
a perpetrator object.
}
\examples{
property_table(examplinib_parent)
property_table(examplinib_compounds)
}
