% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Protocol-class}
\alias{Protocol-class}
\title{Protocol retirve the information of a Protocol used in the experiment}
\description{
Protocol retirve the information of a Protocol used in the experiment
}

