% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PublicationDetail-class}
\alias{PublicationDetail-class}
\title{PublicationDetail, This class contains a Publication Record}
\description{
PublicationDetail, This class contains a Publication Record
}

