% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PublicationResult-class}
\alias{PublicationResult-class}
\title{PiublicationResult, This class contains a list of publication for specific Query}
\description{
PiublicationResult, This class contains a list of publication for specific Query
}

