% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{read_cyp_inducer}
\alias{read_cyp_inducer}
\title{Read csv-formatted CYP inducer data}
\usage{
read_cyp_inducer(source)
}
\arguments{
\item{source}{The connection.}
}
\value{
The inducer data as data frame.
}
\description{
Read csv-formatted CYP inducer data
}
\examples{
read_cyp_inducer(paste0(system.file(package="ddir"), "/tests/testthat/fixtures/cyp-induction.csv"))
}
