% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{read_cyp_inhibitor_data}
\alias{read_cyp_inhibitor_data}
\title{Read CYP inhibition data}
\usage{
read_cyp_inhibitor_data(source)
}
\arguments{
\item{source}{The file or text connection to read from.}
}
\value{
A data frame.
}
\description{
Read CYP inhibition data from a file or text connection.
}
\details{
The following, comma-separated fields are expected in the input (in this
order):
\itemize{
\item 'name' The perpetrator compound name as character.
\item 'cyp' The CYP enzyme as (upper case) character.
\item 'ki' The \eqn{k_i} in µM as numeric.
\item 'source' Optional source information as character.
}

Lines starting with '#' are considered comments and are not evaluated.

The following is an example of a valid input:

\if{html}{\out{<div class="sourceCode">}}\preformatted{    # PARENT
    # compound, CYP, ki, source
    examplinib, CYP1A2,  NA
    examplinib, CYP2B6,  NA,
    examplinib, CYP2C8,  11,   study 001
    examplinib, CYP2C9,  0.6, study 001
    examplinib, CYP2C19, 0.25,   study 001
    examplinib, CYP2D6,  NA,
    examplinib, CYP3A4,  12.5, study 001

    # METABOLITE
    M1,         CYP2C9,  4.4,  study 002
}\if{html}{\out{</div>}}
}
\examples{
read_cyp_inhibitor_data(textConnection(examplinib_cyp_inhibition_string))
}
