% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{read_dmpk}
\alias{read_dmpk}
\title{Read csv-formatted DMPK data}
\usage{
read_dmpk(source)
}
\arguments{
\item{source}{The connection to read from.}
}
\value{
The data as data frame.
}
\description{
Read csv-formatted DMPK data
}
\examples{
read_dmpk(paste0(system.file(package="ddir"), "/tests/testthat/fixtures/cyp-inhibition.csv"))

}
