% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{read_inhibitor_data}
\alias{read_inhibitor_data}
\title{Read csv-formatted DMPK data}
\usage{
read_inhibitor_data(source)
}
\arguments{
\item{source}{The connection to read from.}
}
\value{
The data as data frame.
}
\description{
Read csv-formatted DMPK data
}
