% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{read_perpetrators}
\alias{read_perpetrators}
\title{Read perpetrator information from file or text string}
\usage{
read_perpetrators(source)
}
\arguments{
\item{source}{The connection to read from.}
}
\value{
A list of perpetrator objects.
}
\description{
Read perpetrator information from file or text string
}
\examples{
read_perpetrators(paste0(system.file(package="ddir"), "/tests/testthat/fixtures/compounds.csv"))
}
