% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyp.R
\name{read_string_dmpk_data}
\alias{read_string_dmpk_data}
\title{Read csv-formatted DMPK data from string}
\usage{
read_string_dmpk_data(x)
}
\arguments{
\item{x}{The string containing the data in csv format.}
}
\value{
The data as data frame.
}
\description{
Read csv-formatted DMPK data from string
}
