% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{read_string_perpetrators}
\alias{read_string_perpetrators}
\title{Load perpetrator data from csv-formatted string}
\usage{
read_string_perpetrators(input_string)
}
\arguments{
\item{input_string}{The string variable holding the csv-formatted data.}
}
\value{
A list of perpetrator objects
}
\description{
Load perpetrator data from csv-formatted string
}
\seealso{
[load_perpetrators()]
}
