% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{read_tdi_data}
\alias{read_tdi_data}
\title{Read csv-formatted CYP TDI data}
\usage{
read_tdi_data(source)
}
\arguments{
\item{source}{The connection to read from.}
}
\value{
A data frame.
}
\description{
This function reads comma-separated CYP TDI data from a file or text
connection.
}
\details{
The following fields are expected (in this order):
\itemize{
\item 'name' The perpetrator compound name as character.
\item 'cyp' The CYP enzyme as (upper case) character.
\item 'ki' The \eqn{K_I} in µM as numeric.
\item 'kinact' The \eqn{k_{inact}} in \eqn{1/h} as numeric.
\item 'source' Optional source information as character.
Lines starting with '#' are interpreted as comments and are not evaluated.
}

A valid data set is, e.g.,

\if{html}{\out{<div class="sourceCode">}}\preformatted{
# compound, CYP,    ki,   kinact, source
examplinib, CYP3A4, 0.17, 0.04, study 001
}\if{html}{\out{</div>}}
}
