% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset.R
\name{search.DatasetsSummary}
\alias{search.DatasetsSummary}
\title{search.DatasetsSummary}
\usage{
search.DatasetsSummary(query = "", start = 0, size = 20, faceCount = 20)
}
\arguments{
\item{query}{the current query of the search}

\item{start}{where to start the page}

\item{size}{how many elements to read}

\item{faceCount}{the information of all the facets for the search}

\item{sort.field}{the field to sort the results}
}
\value{
DatasetResults
}
\description{
This function search for dataset summaries in the resource
}

