% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perpetrator.R
\name{show,compound-method}
\alias{show,compound-method}
\title{Print method for compound objects.}
\usage{
\S4method{show}{compound}(object)
}
\arguments{
\item{compound}{Compound object.}
}
\value{
Nothing.
}
\description{
Print method for compound objects.
}
