% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{show,ddi_risk-method}
\alias{show,ddi_risk-method}
\title{Show ddi_risk object}
\usage{
\S4method{show}{ddi_risk}(object)
}
\arguments{
\item{object}{A \code{ddi_risk} object.}
}
\value{
Nothing.
}
\description{
Show ddi_risk object
}
