% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_vitro_inducer.R
\name{show,induction_experiment-method}
\alias{show,induction_experiment-method}
\title{Show method for induction_experiment objects}
\usage{
\S4method{show}{induction_experiment}(object)
}
\arguments{
\item{object}{An \code{induction_experiment} object.}
}
\value{
Nothing.
}
\description{
Show method for induction_experiment objects
}
