% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.varlabel.R
\name{show.varlabel}
\alias{show.varlabel}
\title{Show label information of a variable}
\usage{
show.varlabel(data = data, var.name = var.name)
}
\arguments{
\item{data}{a data frame object created with read.codebook()}

\item{var.name}{single text object of a varialbe name (could be non-quoted)}
}
\value{
This function returns a label of a variable, which are retrieved from '//dataDscr//var@name' in a ddi meta data file.
}
\description{
Show label information of a variable
}
\examples{
ddi.file.url  <- system.file("extdata","samplesurvey2018_ddi.xml",package="DDIR")
data.file.url <- system.file("extdata","samplesurvey2018_data.csv",package="DDIR")

read.codebook(ddi.file.url, data.file.url) -> dataset

show.question(dataset,'Q2')
show.question(dataset$Q2)

require(tidyverse)
dataset \%>\% show.varlabel(Q2)
dataset \%>\% show.varlabel('Q2')

}
