% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_question.R
\name{show_question}
\alias{show_question}
\title{Show question sentenses of a variable}
\usage{
show_question(data = data, var.name = var.name)
}
\arguments{
\item{data}{a data frame object created with read.codebook()}

\item{var.name}{single text object of a varialbe name (could be non-quoted)}
}
\value{
This function returns a question sentense of the varialbe, which is retrieved from '//dataDscr//var//qstn//text()' in a ddi meta data file.
}
\description{
Show question sentenses of a variable
}
\examples{
ddi.file.url  <- system.file("extdata","samplesurvey2018_ddi.xml",package="DDIR")
data.file.url <- system.file("extdata","samplesurvey2018_data.csv",package="DDIR")

read_codebook(ddi.file.url, data.file.url) -> dataset

show_question(dataset,'Q1')
show_question(dataset$Q1)

require(tidyverse)
dataset \%>\% show_question(Q1)
dataset \%>\% show_question('Q1')

}
