% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Similar-class}
\alias{Similar-class}
\title{Similar, This class contains the information of one particuar dataset in OmicsDI}
\description{
Similar, This class contains the information of one particuar dataset in OmicsDI
}

