% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyp.R
\name{static_cyp_induction_risk.list}
\alias{static_cyp_induction_risk.list}
\title{Basic static CYP induction risk}
\usage{
\method{static_cyp_induction_risk}{list}(perp, cyp_ind)
}
\arguments{
\item{perp}{The perpetrator object or a list thereof.}

\item{cyp_ind}{The CYP induction data as data frame.}
}
\value{
A data frame.
}
\description{
Basic static CYP induction risk
}
\examples{
static_cyp_induction_risk(examplinib_compounds, examplinib_cyp_induction_data)
}
