% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyp.R
\name{static_cyp_induction_risk_table.list}
\alias{static_cyp_induction_risk_table.list}
\title{Tables of the basic static CYP induction risk}
\usage{
\method{static_cyp_induction_risk_table}{list}(perp, cyp_ind, na.rm = F)
}
\arguments{
\item{perp}{The list of perpetrator objects.}

\item{cyp_ind}{The CYP induction data as data frame.}

\item{na.rm}{Remove rows with lacking ki data (i.e., where ki == NA).}
}
\value{
A list of markdown-formatted tables.
}
\description{
Tables of the basic static CYP induction risk
}
\examples{
static_cyp_induction_risk_table(examplinib_compounds, examplinib_cyp_induction_data)
}
