% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{StatRecord-class}
\alias{StatRecord-class}
\title{StatRecord, This class contains the information of statistics records}
\description{
StatRecord, This class contains the information of statistics records
}

