% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdi_tools.R
\name{tdimod}
\alias{tdimod}
\title{Derive TDI parameters from enzyme activity data}
\usage{
tdimod(x, object = NULL)
}
\arguments{
\item{x}{In vitro data as data frame. The following columns are expected:
\itemize{
\item TIME Duration of Pre-incubation with the precipitant and NADPH in minutes
\item CONC Precipitant concentration in uM
\item ACT Enzyme activity in percent
}}

\item{object}{The DDI target object as character or NULL.}
}
\value{
A list with the following items:
\itemize{
\item data The input data with linear models of ACT over TIME by CONC.
\item kobs The kobs parameters by precipitant concentration.
\item kobs_plot A ggplot object showing the kobs fit to the ACT data.
\item tdi_param The TDI parameters, Kinact and KI, derived from an Emax model of
kobs over CONC.
\item tdi_plot A ggplot object showing the Emnax modeling of kobs over CONC.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
kobs is fitted to the enzyme activity data as a first-order process for each
inhibitor concentration.
kinact (the maximal inhibition rate) and KI (inhibitor concentration at the
half-maximal kobs) are fitted from kobs over inhibitor concentration using a
Emax model.
}
\examples{
tdimod(examplinib_in_vitro_tdi)
}
