% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statements.R
\name{transporter_inhibition_risk_summary.perpetrator}
\alias{transporter_inhibition_risk_summary.perpetrator}
\title{Title}
\usage{
\method{transporter_inhibition_risk_summary}{perpetrator}(
  perp,
  transporter_inh,
  transporter_ref = transporter_reference_data
)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{transporter_inh}{Transporter inhibition data as data frame. The
following fields are expected:
\itemize{
\item 'name' The perpetrator compound name.
\item 'cyp' The UGT enzyme as (upper case) character.
\item 'ic50' The \eqn{IC_{50}} of the inhibition in μM.
\item 'source' Optional source information as character.
}}

\item{transporter_ref}{Transporter reference data, see
\link{transporter_reference_data} for details.}
}
\value{
Text output
}
\description{
Title
}
