% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transporter.R
\name{transporter_inhibition_risk_table}
\alias{transporter_inhibition_risk_table}
\title{Table of drug transporter inhibition risks}
\usage{
transporter_inhibition_risk_table(
  perp,
  transporter_ic50,
  transporter_ref = transporter_reference_data,
  na.rm = F
)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{transporter_ic50}{Transporter inhibition IC50 as data frame.}

\item{transporter_ref}{Transporter reference data.}

\item{na.rm}{Remove rows with lacking ki data (i.e., where ki == NA).}
}
\value{
A markdown-formatted table.
}
\description{
Table of drug transporter inhibition risks
}
