% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statements.R
\name{ugt_inhibition_risk_summary.list}
\alias{ugt_inhibition_risk_summary.list}
\title{Summary statement for UGT inhibition for a list of compounds}
\usage{
\method{ugt_inhibition_risk_summary}{list}(perp, ...)
}
\arguments{
\item{perp}{A perpetrator object.}

\item{...}{Further parameters.}
}
\value{
Markdown-formatted text output.
}
\description{
Summary statement for UGT inhibition for a list of compounds
}
