% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{basic_cyp_inhibition_risk}
\alias{basic_cyp_inhibition_risk}
\title{Basic CYP inhibition risk}
\usage{
basic_cyp_inhibition_risk(perp, cyp_inh)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{cyp_inh}{CYP inhibition data object.}
}
\value{
DDI risk object.
}
\description{
Evaluate the clinical risk for direct (reversible) CYP inhibition according
to the basic model defined in the \href{https://www.ema.europa.eu/en/documents/scientific-guideline/ich-m12-guideline-drug-interaction-studies-step-5_en.pdf}{ICH M12 guideline}.
}
\details{
For the basic modeling of direct (reversible) CYP enzyme inhibition,
the ratio of the relevant inhibitor concentration to the \eqn{K_i} of the
respective CYP enzyme is considered, i.e., \eqn{R} for hepatic enzymes and
\eqn{R_{gut}} for intestinal enzymes (refer to Section 2.1.2.1 of the \href{https://www.ema.europa.eu/en/documents/scientific-guideline/ich-m12-guideline-drug-interaction-studies-step-5_en.pdf}{ICH M12 guidance document}).
\subsection{Liver}{

\deqn{R=\frac{C_{max,ss,u}}{K_{i,u}}}

\eqn{R} values > 0.02, i.e., maximal unbound perpetrator concentrations
50-fold over \eqn{K_i} are considered to indicate a potential clinical CYP
inhibition risk using this method.
}

\subsection{Intestine}{

\deqn{R_{gut}=\frac{I_{gut}}{K_{i,u}}}

where

\deqn{I_{gut}=\frac{Dose}{250\ mg}}

\eqn{R} values > 10 are considered to indicate a clinical risk for
intestinal CYP3A inhibition.

In the output, the columns \code{risk_hep} and \code{risk_intest} indicate whether
the regulatory threshold is reached.
}
}
