% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyp.R
\name{basic_cyp_inhibition_risk}
\alias{basic_cyp_inhibition_risk}
\title{Generic function to evaluate the basic CYP inhibition risk}
\usage{
basic_cyp_inhibition_risk(perp, cyp_inh)
}
\arguments{
\item{perp}{The perpetrator as perpetrator object or list thereof.}

\item{cyp_inh}{The CYP inhibitor data as data frame.}
}
\value{
A data frame.
}
\description{
Generic function to evaluate the basic CYP inhibition risk
}
