% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyp.R
\name{basic_cyp_inhibition_risk}
\alias{basic_cyp_inhibition_risk}
\title{Basic CYP inhibition parameters}
\usage{
basic_cyp_inhibition_risk(perp, cyp_inh)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{cyp_inh}{CYP inhibiton data as data frame.}
}
\value{
Data frame.
}
\description{
Basic CYP inhibition parameters
}
