% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyp.R
\name{basic_cyp_inhibition_risk_table}
\alias{basic_cyp_inhibition_risk_table}
\title{Basic CYP inhibition risk table}
\usage{
basic_cyp_inhibition_risk_table(
  perp,
  cyp_inh,
  na.rm = FALSE,
  show_dose = FALSE
)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{cyp_inh}{}

\item{na.rm}{Switch to define whether rows with lacking \eqn{K_i} data are
removed from the output (i.e., where \code{ki == NA}). Defaults to \code{FALSE}.}

\item{show_dose}{Show perpetrator dose in table title, defaults to \code{FALSE.}}
}
\value{
A markdown-formatted table, or an empty string.
}
\description{
This function generates a markdown-formatted table of the direct (reversible)
CYP inhibition risk assessment. See \code{\link[=basic_cyp_inhibition_risk]{basic_cyp_inhibition_risk()}} for details
on the calculation of the risk.
}
\examples{
basic_cyp_inhibition_risk_table(examplinib_parent,
  examplinib_cyp_inhibition_data)
basic_cyp_inhibition_risk_table(examplinib_compounds,
  examplinib_cyp_inhibition_data, na.rm = TRUE)
basic_cyp_inhibition_risk_table(examplinib_compounds,
  examplinib_cyp_inhibition_data, na.rm = TRUE, show_dose = TRUE)
}
\seealso{
\code{\link[=basic_cyp_inhibition_risk]{basic_cyp_inhibition_risk()}}
}
