% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyp.R
\name{basic_cyp_inhibition_risk_table}
\alias{basic_cyp_inhibition_risk_table}
\title{Generic function for basic CYP inhibition risk evaluation as table}
\usage{
basic_cyp_inhibition_risk_table(perp, cyp_inh, na.rm = F)
}
\arguments{
\item{perp}{The perpetrator as perpetrator object or list thereof.}

\item{cyp_inh}{CYP inhibition data as data frame.}

\item{na.rm}{Remove rows with lacking ki data (i.e., where ki == NA).}
}
\description{
Generic function for basic CYP inhibition risk evaluation as table
}
\examples{
basic_cyp_inhibition_risk_table(examplinib_parent, examplinib_cyp_inhibition_data)
}
