% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{basic_cyp_tdi_risk}
\alias{basic_cyp_tdi_risk}
\title{Title}
\usage{
basic_cyp_tdi_risk(perp, cyp_tdi, cyp_kdeg = cyp_turnover)
}
\arguments{
\item{perp}{Perpetrator object.}

\item{cyp_tdi}{Inhibitor object.}

\item{cyp_kdeg}{Data frame}
}
\value{
DDI risk object
}
\description{
Title
}
