% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_assessment.R
\name{basic_cyp_tdi_risk}
\alias{basic_cyp_tdi_risk}
\title{CYP time-dependent inhibition risk}
\usage{
basic_cyp_tdi_risk(perp, cyp_tdi, cyp_kdeg = cyp_turnover)
}
\arguments{
\item{perp}{Perpetrator object.}

\item{cyp_tdi}{inhibition_data object.}

\item{cyp_kdeg}{CYP turnover rates as data frame, defaults to
\link{cyp_turnover}.}
}
\value{
DDI risk object.
}
\description{
CYP time-dependent inhibition risk
}
\examples{
basic_cyp_tdi_risk(examplinib, examplinib_cyp_tdi)
}
