% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyp.R
\name{basic_cyp_tdi_risk}
\alias{basic_cyp_tdi_risk}
\title{Basic modeling of the CYP time-dependent inhibition risk}
\usage{
basic_cyp_tdi_risk(perp, tdi, cyp_kdeg = cyp_turnover)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{tdi}{The CYP TDI data as data frame.}

\item{cyp_kdeg}{The CYP turnover data as data frame. Defaults to the
built-in reference data.}
}
\value{
A data frame.
}
\description{
In the present version, only the risk for hepatic TDI for CYP enzymes is
calculated.
}
\examples{
basic_cyp_tdi_risk(examplinib_parent, examplinib_cyp_tdi_data)
}
\seealso{
\link{cyp_turnover}
}
