% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyp.R
\name{basic_cyp_tdi_risk_table}
\alias{basic_cyp_tdi_risk_table}
\title{Basic CYP time-dependent inhibition risk table}
\usage{
basic_cyp_tdi_risk_table(perp, tdi, cyp_kdeg = cyp_turnover, na.rm = TRUE)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{tdi}{The CYP TDI data as data frame.}

\item{cyp_kdeg}{The CYP turnover data as data frame. Defaults to the
built-in reference data.}

\item{na.rm}{Boolean to define whether rows with lacking \eqn{K_I} or
\eqn{k_{deg}} data are removed from the output.}
}
\value{
A markdown-formatted table.
}
\description{
Basic CYP time-dependent inhibition risk table
}
\examples{
basic_cyp_tdi_risk_table(examplinib_parent, examplinib_cyp_tdi_data)
}
