% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ugt.R
\name{basic_ugt_inhibition_risk}
\alias{basic_ugt_inhibition_risk}
\title{UGT inhibition risk}
\usage{
basic_ugt_inhibition_risk(perp, ugt_inh)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{ugt_inh}{UGT inhibition data as data frame, with 'value' representing
the respective \eqn{IC_{50}}.}
}
\value{
A markdown-formatted table.
}
\description{
This function evaluates the clinical risk for reversible inhibition of UGT
enzymes according to the relevant
\href{https://www.fda.gov/media/134582/download}{FDA} and
\href{https://www.ema.europa.eu/en/documents/scientific-guideline/guideline-investigation-drug-interactions-revision-1_en.pdf}{EMA}
guidelines.
}
\details{
This function assumes that the UGT inhibition data is provided as
\eqn{IC_{50}}. According to
\href{https://doi.org/10.1016/0006-2952(73)90196-2}{Cheng, Prusoff 1973}),
\eqn{K_i} can be assumed to be \eqn{IC_{50}/2} at the experimental conditions
commonly used in the in vitro ihibition studies where substrate
concentrations close to \eqn{K_M} are used.

The relevant metric for basic modeling of the UGT inhibition risk is
\eqn{R_1=I_{max,ss,u}/K_{i,u}}. For the clinical risk assessment, a cut-off
of \eqn{R<1.02} applies.

Refer to the documentation to the \code{\link[=key_concentrations]{key_concentrations()}} function for details
on the calculation of \eqn{I_{max,ss,u}}.
}
\seealso{
\code{\link[=key_concentrations]{key_concentrations()}}

\code{\link[=basic_ugt_inhibition_risk_table]{basic_ugt_inhibition_risk_table()}}
}
