% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ugt.R
\name{basic_ugt_inhibition_risk}
\alias{basic_ugt_inhibition_risk}
\title{UGT inhibition risk}
\usage{
basic_ugt_inhibition_risk(perp, ugt_inh)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{ugt_inh}{UGT inhibiton data as data frame, with value repersenting the
IC50 (!).}
}
\value{
A markdown-formatted table.
}
\description{
This function assumes that the UGT inhibition data is provided as ICC50.
According to Cheng-Prusoff, Ki can be assumed to be IC50/2 at the
usual experimental conditions where substrate concentrations close to Km
are used.
}
