% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ugt.R
\name{basic_ugt_inhibition_risk_table}
\alias{basic_ugt_inhibition_risk_table}
\title{UGT inhibition risk table}
\usage{
basic_ugt_inhibition_risk_table(perp, ugt_inh, na.rm = F, show_dose = FALSE)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{ugt_inh}{UGT inhibition data as data frame, The following fields are
expected:
\itemize{
\item 'name' The name of the perpetrator compound.
\item 'ugt' The UGT enzyme as (upper case) character.
\item 'ic50' The \eqn{IC_{50}} in \eqn{\mu M} as numeric.
\item 'source' Optional source information as character.
}}

\item{na.rm}{Boolean to define whether rows with lacking \eqn{K_i} data are
removed from the output (i.e., where \code{ki == NA}).}

\item{show_dose}{Show_dose Show perpetrator dose in table title, defaults to
\code{FALSE.}}
}
\value{
A markdown-formatted table.
}
\description{
This function generates a markdown-formatted table of the reversible
UGT inhibition risk assessment. See \code{\link[=basic_ugt_inhibition_risk]{basic_ugt_inhibition_risk()}} for details
on the calculation of the risk.
}
\examples{
basic_ugt_inhibition_risk_table(examplinib_parent,
  examplinib_ugt_inhibition_data)
basic_ugt_inhibition_risk_table(examplinib_compounds,
  examplinib_ugt_inhibition_data)
basic_ugt_inhibition_risk_table(examplinib_compounds,
  examplinib_ugt_inhibition_data, show_dose = TRUE)
}
\seealso{
\code{\link[=basic_ugt_inhibition_risk]{basic_ugt_inhibition_risk()}}
}
