% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ugt.R
\name{basic_ugt_inhibition_risk_table}
\alias{basic_ugt_inhibition_risk_table}
\title{UGT inhibition risk table}
\usage{
basic_ugt_inhibition_risk_table(perp, ugt_inh, na.rm = F)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{ugt_inh}{UGT inhibition data as data frame, with 'value' representing
the respective \eqn{IC_{50}}. Refer to \link{examplinib_ugt_inhibition_data} for
a sample UGT inhibition data set.}

\item{na.rm}{Boolean to define whether rows with lacking \eqn{K_i} data are
removed from the output (i.e., where \code{ki == NA}).}
}
\value{
A markdown-formatted table.
}
\description{
This function generates a markdown-formatted table of the reversible
UGT inhibition risk assessment. See \code{\link[=basic_ugt_inhibition_risk]{basic_ugt_inhibition_risk()}} for details
on the calculation of the risk.
}
\seealso{
\code{\link[=basic_ugt_inhibition_risk]{basic_ugt_inhibition_risk()}}
}
