% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ugt.R
\name{basic_ugt_inhibition_risk_table}
\alias{basic_ugt_inhibition_risk_table}
\title{UGT inhibition risk table}
\usage{
basic_ugt_inhibition_risk_table(perp, ugt_inh, na.rm = F)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{ugt_inh}{UGT inhibition data as data frame, with 'value' representing
the IC50 (!).}

\item{na.rm}{Remove rows with lacking ki data (i.e., where ki == NA).}
}
\value{
A markdown-formatted table.
}
\description{
This function generates a markdown-formatted table of the reversible
UGT inhibition risk assessment. See \code{\link[=basic_ugt_inhibition_risk]{basic_ugt_inhibition_risk()}} for details
on the calculation of the risk.
}
\seealso{
\code{\link[=basic_ugt_inhibition_risk]{basic_ugt_inhibition_risk()}}
}
