% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{conc_table.list}
\alias{conc_table.list}
\title{Table of key perpetrator concentrations for a list of compounds}
\usage{
\method{conc_table}{list}(perp)
}
\arguments{
\item{perp}{The compounds as list of compound objects.}
}
\description{
Table of key perpetrator concentrations for a list of compounds
}
