% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{conc_table.perpetrator}
\alias{conc_table.perpetrator}
\title{Key perpetrator concentrations}
\usage{
\method{conc_table}{perpetrator}(perp)
}
\arguments{
\item{perp}{The perpetrator object.}
}
\value{
A markdown-formatted table.
}
\description{
This function generates a markdown-formatted table of the key perpetrator
concentrations used for the assessment of the DDI perpetrator potential. See
\code{\link[=key_concentrations]{key_concentrations()}} for details on their calculation.
}
\examples{
conc_table(examplinib_parent)
}
\seealso{
\code{\link[=key_concentrations]{key_concentrations()}}
}
