% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{conc_table}
\alias{conc_table}
\title{Table of key perpetrator concentrations}
\usage{
conc_table(perp)
}
\arguments{
\item{perp}{The perpetrator object.}
}
\value{
Key perpetrator concentrations as a markdown-formatted table.
}
\description{
Table of key perpetrator concentrations
}
