% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{conc_table}
\alias{conc_table}
\title{Key perpetrator concentrations}
\usage{
conc_table(perp)
}
\arguments{
\item{perp}{The perpetrator object or a list of perpetrator objects.}
}
\value{
A markdown-formatted table.
}
\description{
This function generates a markdown-formatted table of the key
concentrations used for the assessment of the DDI perpetrator potential. See
\code{\link[=key_concentrations]{key_concentrations()}} for details on the calculation of the concentrations.
}
\examples{
conc_table(examplinib_parent)
conc_table(examplinib_compounds)
}
\seealso{
\code{\link[=key_concentrations]{key_concentrations()}}
}
