% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{conc_table}
\alias{conc_table}
\title{Generic function to display ky perpetrator concentrations}
\usage{
conc_table(perp)
}
\arguments{
\item{perp}{The object (compund object or list thereof)}
}
\description{
Generic function to display ky perpetrator concentrations
}
