% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{conc_table}
\alias{conc_table}
\title{Table of key perpetrator concentrations}
\usage{
conc_table(perp, show_dose = FALSE)
}
\arguments{
\item{perp}{The perpetrator object or a list of perpetrator objects.}

\item{show_dose}{Show perpetrator dose in table title, defaults to \code{FALSE.}}
}
\value{
A markdown-formatted table.
}
\description{
This function generates a markdown-formatted table of the key perpetrator
concentrations used for the assessment of the DDI perpetrator potential. See
\code{\link[=key_concentrations]{key_concentrations()}} for details on their calculation.
}
\examples{
conc_table(examplinib_parent)
conc_table(examplinib_compounds)
conc_table(examplinib_compounds, show_dose = TRUE)
}
\seealso{
\code{\link[=key_concentrations]{key_concentrations()}}
}
