% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statements.R
\name{cyp_inhibition_risk_summary.list}
\alias{cyp_inhibition_risk_summary.list}
\title{Summary statement for CYP inhibition for a list of perpetrators}
\usage{
\method{cyp_inhibition_risk_summary}{list}(perp, ...)
}
\arguments{
\item{perp}{Perpetrator object.}

\item{...}{Further parameters.}
}
\value{
Text output.
}
\description{
Summary statement for CYP inhibition for a list of perpetrators
}
\examples{
cyp_inhibition_risk_summary.list(examplinib_compounds, examplinib_cyp_inhibition_data)
}
