% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statements.R
\name{cyp_inhibition_risk_summary}
\alias{cyp_inhibition_risk_summary}
\title{Summary statement for CYP inhibition}
\usage{
cyp_inhibition_risk_summary(perp, cyp_inh)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{cyp_inh}{CYP inhibition data as data frame. The following fields are
expected:
\itemize{
\item 'name' The name of the perpetrator compound.
\item 'cyp' The CYP enzyme as (upper case) character.
\item 'ki' The \eqn{K_i} in \eqn{\mu M} as numeric.
\item 'source' Optional source information as character.
}}
}
\value{
Markdown-formatted text output.
}
\description{
Summary statement for CYP inhibition
}
\examples{
cyp_inhibition_risk_summary(examplinib_parent, examplinib_cyp_inhibition_data)
cyp_inhibition_risk_summary(examplinib_compounds, examplinib_cyp_inhibition_data)
}
