% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cyp_reference_substrates}
\alias{cyp_reference_substrates}
\title{CYP reference substrate data}
\format{
## `cyp_reference_substrates`
}
\source{
FDA and EMA guidelines.
}
\usage{
cyp_reference_substrates
}
\description{
CYP reference substrate data
}
\keyword{datasets}
