% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cyp_reference_substrates}
\alias{cyp_reference_substrates}
\title{CYP reference substrate data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 5 columns.
}
\source{
FDA and EMA guidelines.
}
\usage{
cyp_reference_substrates
}
\description{
CYP reference substrates commonly used in the mechanistic static assessment
of the CYP DDI potential of drugs.
}
\details{
The CYP reference substrates currently implemented include:
\preformatted{
cyp     , substrate   , fgut , fm   , fmcyp
CYP1A2  , tizanidine  , 1    , 0.95 , 0.98
CYP2C8  , repaglinide , 1    , 1    , 0.61
CYP2C9  , S-warfarin  , 1    , 1    , 0.91
CYP2C19 , omeprazole  , 1    , 1    , 0.87
CYP3A4  , midazolam   , 0.57 , 0.96 , 1
}
}
\seealso{
\code{\link[=mech_stat_cyp_risk]{mech_stat_cyp_risk()}}

\link{mech_stat_cyp_risk_table}
}
\keyword{datasets}
