% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statements.R
\name{cyp_tdi_risk_summary.list}
\alias{cyp_tdi_risk_summary.list}
\title{Summary statement for CYP TDI for a list of compounds}
\usage{
\method{cyp_tdi_risk_summary}{list}(perp, ...)
}
\arguments{
\item{perp}{A list of perpetrator objects.}

\item{...}{Further parameters.}
}
\value{
Markdown-formatted text output.
}
\description{
Summary statement for CYP TDI for a list of compounds
}
\examples{
cyp_tdi_risk_summary(examplinib_compounds, examplinib_cyp_tdi_data)
}
