% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statements.R
\name{cyp_tdi_risk_summary.perpetrator}
\alias{cyp_tdi_risk_summary.perpetrator}
\title{Summary statement for CYP TDI}
\usage{
\method{cyp_tdi_risk_summary}{perpetrator}(perp, cyp_tdi, cyp_kdeg = cyp_turnover)
}
\arguments{
\item{perp}{The perpetrator object.}
}
\value{
Text output.
}
\description{
Summary statement for CYP TDI
}
\examples{
cyp_tdi_risk_summary.perpetrator(examplinib_parent, examplinib_cyp_tdi_data)
}
