% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{database,DatasetDetail-method}
\alias{database,DatasetDetail-method}
\title{Returns an acession id}
\usage{
\S4method{database}{DatasetDetail}(object)
}
\arguments{
\item{object}{a DatasetDetail}
}
\value{
database accession
}
\description{
Returns an acession id
}
\author{
Yasset Perez-Riverol
}
