% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\docType{class}
\name{ddi_risk-class}
\alias{ddi_risk-class}
\title{Generic DDI risk class definition}
\value{
An ddi_risk class object.
}
\description{
Generic DDI risk class definition
}
\section{Slots}{

\describe{
\item{\code{precipitant}}{Name of perpetrator.}

\item{\code{table}}{Tabulated risk assessment data.}
}}

