% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statements.R
\name{ddi_risk_summary}
\alias{ddi_risk_summary}
\title{Overall DDI risk summary text block}
\usage{
ddi_risk_summary(
  perp,
  cyp_inh = NULL,
  cyp_tdi = NULL,
  cyp_ind = NULL,
  ugt_inh = NULL,
  transporter_inh = NULL,
  d = 1,
  include_induction = TRUE,
  substr = cyp_reference_substrates,
  cyp_kdeg = cyp_turnover
)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{cyp_inh}{CYP inhibition data object.}

\item{cyp_ind}{}

\item{ugt_inh}{UGT inhibition data object.}

\item{transporter_inh}{Transporter inhibition data as data frame. The
following fields are expected:
\itemize{
\item 'name' The perpetrator compound name.
\item 'cyp' The UGT enzyme as (upper case) character.
\item 'ic50' The \eqn{IC_{50}} of the inhibition in μM.
\item 'source' Optional source information as character.
}}

\item{d}{}

\item{cyp_kdeg}{}
}
\value{
Markdown-formatted text output.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
ddi_risk_summary(examplinib_compounds, examplinib_cyp_inhibition_data,
  examplinib_cyp_tdi_data, examplinib_cyp_induction_data,
  examplinib_ugt_inhibition_data, examplinib_transporter_inhibition_data)
ddi_risk_summary(examplinib_parent, examplinib_cyp_inhibition_data)
}
