% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddir-package.R
\docType{package}
\name{ddir-package}
\alias{ddir}
\alias{ddir-package}
\title{ddir: DDI-Codebook 2.5}
\description{
DDI-Codebook 2.5 implementation in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gorcha/ddir}
  \item Report bugs at \url{https://github.com/gorcha/ddir/issues}
}

}
\author{
\strong{Maintainer}: Danny Smith \email{danny@gorcha.org}

}
\keyword{internal}
