% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddir-package.R
\docType{package}
\name{ddir-package}
\alias{ddir}
\alias{ddir-package}
\title{ddir: Assessment of the clinical DDI risk for drug compounds}
\description{
The goal of ddir is to facilitate the analysis of the drug-drug interaction risk for drug compounds as per the relevant FDA and EMA guidelines.
}
\keyword{internal}
