% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DictWord-class}
\alias{DictWord-class}
\title{DictWord, This class contains the information of a list of Dictionary Words}
\description{
DictWord, This class contains the information of a list of Dictionary Words
}

