% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{dose.perpetrator}
\alias{dose.perpetrator}
\title{Title}
\usage{
\method{dose}{perpetrator}(obj)
}
\arguments{
\item{obj}{Perpetrator object.}
}
\value{
Dose as character.
}
\description{
Title
}
\examples{
dose(examplinib_parent)
dose(examplinib_metabolite)
}
