% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_compounds}
\alias{examplinib_compounds}
\title{Fictional substrate data}
\format{
An object of class \code{list} of length 2.
}
\source{
Fictional data, made up for demo purposes.
}
\usage{
examplinib_compounds
}
\description{
A list of two: \code{perpetrator} objects:
\itemize{
\item examplinib, a fictional drug.
\item M1, a fictional metabolite of examplinib.
}
}
\keyword{datasets}
