% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_compounds_string}
\alias{examplinib_compounds_string}
\title{Examplinib compound data as string}
\format{
An object of class \code{character} of length 1.
}
\source{
Fictional data, made up for demo purposes.
}
\usage{
examplinib_compounds_string
}
\description{
A character string containing data for two perpetrator objects:
\itemize{
\item examplinib, a fictional drug.
\item M1, a fictional metabolite of examplinib.
}
}
\seealso{
\code{\link[=read_perpetrators]{read_perpetrators()}}
}
\keyword{datasets}
