% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_cyp_induction_data}
\alias{examplinib_cyp_induction_data}
\title{Fictional CYP induction data}
\format{
An object of class \code{data.frame} with 14 rows and 6 columns.
}
\source{
Fictional data, made up for demo purposes.
}
\usage{
examplinib_cyp_induction_data
}
\description{
Fictional CYP induction data
}
\keyword{datasets}
