% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_cyp_tdi}
\alias{examplinib_cyp_tdi}
\title{Examplinib CYP TDI data}
\format{
An object of class \code{inhibition_data} (inherits from \code{data.frame}) with 1 rows and 4 columns.
}
\source{
Fictional data, made up for demo purposes.
}
\usage{
examplinib_cyp_tdi
}
\description{
Examplinib CYP TDI data
}
\details{
\preformatted{
      name    cyp   ki kinact    source
examplinib CYP3A4 30.7   0.04 study 001
}
}
\keyword{datasets}
