% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_in_vitro_ind}
\alias{examplinib_in_vitro_ind}
\title{in vitro sample data for CYP3A4 mRNA induction}
\format{
A induction_experiment object frame with 7 columns:
\itemize{
\item DONOR Hepatocyte donor
\item TYPE Test  or positive control
\item CONC Precipitant concentration
\item OBJECT DDI target
\item FOLD mRNA fold change
\item REL mRNA induction in percent relative to positive control
\item SOURCE Source study report
}
}
\source{
Fully synthetic data set to demonstrate CYP mRNA induction data
evaluation.
}
\usage{
examplinib_in_vitro_ind
}
\description{
CYP3A4 mRNA fold-change data at different precipitant concentrations
}
\keyword{datasets}
