% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_metabolite}
\alias{examplinib_metabolite}
\title{Perpetrator compound data for examplinib}
\format{
An object of class \code{perpetrator} (inherits from \code{data.frame}) with 12 rows and 3 columns.
}
\source{
Fictional data, made up for demo purposes.
}
\usage{
examplinib_metabolite
}
\description{
Perpetrator compound data for examplinib
}
\details{
\preformatted{
                param  value    source
name             name     M1
oral             oral  FALSE
mw                 mw 506.56
dose             dose     NA
imaxss         imaxss   1038 study 001
fu                 fu  0.012 study 002
fumic           fumic      1   default
rb                 rb      1 study 002
fa                 fa     NA
fg                 fg     NA
ka                 ka     NA
solubility solubility    Inf   default
}
}
\keyword{datasets}
