% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_metabolite}
\alias{examplinib_metabolite}
\title{Fictional examplinib metabolite compound data}
\format{
An object of class \code{perpetrator} (inherits from \code{data.frame}) with 12 rows and 3 columns.
}
\source{
Fictional data, made up for demo purposes.
}
\usage{
examplinib_metabolite
}
\description{
A \code{perpetrator} object for 'examplinib-M1'.
}
\keyword{datasets}
