% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_parent}
\alias{examplinib_parent}
\title{Fictional examplnib parent compound data}
\format{
An object of class \code{perpetrator} (inherits from \code{data.frame}) with 13 rows and 3 columns.
}
\source{
Fictional data, made up for demo purposes.
}
\usage{
examplinib_parent
}
\description{
A \code{perpetrator} object for 'examplinib'.
}
\keyword{datasets}
