% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_transporter_inhibition}
\alias{examplinib_transporter_inhibition}
\title{Examplinib transporter inhibition data}
\format{
An object of class \code{inhibition_data} (inherits from \code{data.frame}) with 11 rows and 3 columns.
}
\source{
Fictional data, made up for demo purposes.
}
\usage{
examplinib_transporter_inhibition
}
\description{
Examplinib transporter inhibition data
}
\keyword{datasets}
