% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_transporter_inhibition_data}
\alias{examplinib_transporter_inhibition_data}
\title{Examplinib transporter inhibition data}
\format{
An object of class \code{data.frame} with 11 rows and 4 columns.
}
\source{
Fictional data, made up for demo purposes.
}
\usage{
examplinib_transporter_inhibition_data
}
\description{
Examplinib transporter inhibition data
}
\details{
\preformatted{
      name   param      value    source
examplinib    name examplinib
examplinib     Pgp       0.41 study 005
examplinib    BCRP        1.9 study 005
examplinib    OCT1        2.3 study 006
examplinib OATP1B1        177 study 006
examplinib OATP1B3         35 study 006
examplinib    OAT1        271
examplinib    OAT3        300
examplinib    BSEP       12.8
examplinib    OCT2         67 study 006
examplinib   MATE1        3.6 study 006
examplinib  MATE2k        1.1 study 006
}
}
\keyword{datasets}
