% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplinib_ugt_inhibition_data}
\alias{examplinib_ugt_inhibition_data}
\title{Examplinib UGT inhibition data}
\format{
A data frame with the columns 'name', 'param', 'value' and 'source'
}
\source{
Fictional data, made up for demo purposes.
}
\usage{
examplinib_ugt_inhibition_data
}
\description{
Examplinib UGT inhibition data
}
\details{
\preformatted{
      name   param      value    source
        M1    name         M1
        M1  UGT1A1        1.1 study 009
        M1  UGT1A3        5.8 study 009
        M1  UGT1A4        6.2 study 009
        M1  UGT1A6         15 study 009
        M1  UGT1A9        3.6 study 009
        M1  UGT2B7         15 study 009
        M1 UGT2B15        9.6 study 009
        M1 UGT2B17        2.2 study 009
examplinib    name examplinib
examplinib  UGT1A1         15 study 009
examplinib  UGT1A3         15 study 009
examplinib  UGT1A4         15 study 009
examplinib  UGT1A6         15 study 009
examplinib  UGT1A9        3.8 study 009
examplinib  UGT2B7         15 study 009
examplinib UGT2B15         15 study 009
examplinib UGT2B17        6.1 study 009
}
}
\keyword{datasets}
