% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset.R
\name{from.json.LabMember}
\alias{from.json.LabMember}
\title{from.json.LabMember}
\usage{
from.json.LabMember(json.object)
}
\arguments{
\item{json.object}{object}
}
\value{
LabMember
}
\description{
This function convert objects from json to LabMember
}

