<<<<<<< HEAD
% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset.R
\name{get.MetadataSimilars}
\alias{get.MetadataSimilars}
\title{get.MetadataSimilars}
\usage{
get.MetadataSimilars(database = "", accession = "")
}
\arguments{
\item{database}{}

\item{accession}{}
}
\value{
DatasetResults
}
\description{
Dataset Similars by Metadata
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get.MetadataSimilars}
\alias{get.MetadataSimilars}
\title{get.MetadataSimilars
Returns a Biological Molecules similarity for a Dataset}
\usage{
get.MetadataSimilars(accession, database)
}
\arguments{
\item{accession}{id in the original repository}

\item{database}{the original database where this dataset has been store}
}
\value{
List of similars
}
\description{
get.MetadataSimilars
Returns a Biological Molecules similarity for a Dataset
}
\author{
Yasset Perez-Riverol
>>>>>>> yasset
}

