% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get.MetadataSimilars}
\alias{get.MetadataSimilars}
\title{get.MetadataSimilars}
\usage{
get.MetadataSimilars(accession, database)

get.MetadataSimilars(accession, database)
}
\arguments{
\item{accession}{id in the original repository}

\item{database}{the original database where this dataset has been store}
}
\value{
DatasetResults

List of similars
}
\description{
Dataset Similars by Metadata
}
\author{
Yasset Perez-Riverol
}
