% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perpetrator.R
\name{imaxinletu}
\alias{imaxinletu}
\title{Cmax,u at the hepatic inlet}
\usage{
imaxinletu(x, qh = 1.616, molar = TRUE)
}
\arguments{
\item{x}{A perpetrator compound object.}

\item{qh}{Hepatic blood flow in l/min, defaults to 1.616 l/min.}

\item{molar}{Output in molar concentration (uM). Defaults to FALSE (ng/nl).}
}
\value{
Numeric.
}
\description{
Cmax,u at the hepatic inlet
}
