% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipitant.R
\name{imaxintest}
\alias{imaxintest}
\title{Intestinal Cmax,u}
\usage{
imaxintest(x, qent = 18/60, molar = TRUE)
}
\arguments{
\item{x}{A precipitant compound object.}

\item{qent}{Enteric blood flow in l/min, defaults to 0.3 l/min = 18 l/h.}

\item{molar}{Output in molar concentration (uM). Defaults to FALSE (ng/nl).}
}
\value{
Numeric.
}
\description{
Intestinal Cmax,u
}
