% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipitant.R
\name{imaxssu}
\alias{imaxssu}
\title{Cmax,ss,u}
\usage{
imaxssu(x, molar = TRUE)
}
\arguments{
\item{x}{A precipitant compound object.}

\item{molar}{Output in molar concentration (uM). Defaults to FALSE (ng/nl).}
}
\value{
Numeric.
}
\description{
Cmax,ss,u
}
