% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_tools.R
\name{indmod}
\alias{indmod}
\title{Analyze in vitro CYP induction data}
\usage{
indmod(x, use_emax_obs = TRUE, individual_donors = TRUE)
}
\arguments{
\item{x}{In vitro induction data as induction_experiment object.}

\item{use_emax_obs}{Constrain Emax to the observed Emax (default). If FALSE,
Emax will be fitted.}

\item{individual_donors}{Plot donors on individual panels.}
}
\value{
A list with the elements:
\itemize{
\item data The original data, model and model parameters for each fit.
\item fold_plot A ggplot object showing the original data and the fits.
\item ind_param The ec50, hill (n) and emax parameters by donor and DDI object.
\item inducer An inducer object representing the model parameters from the donor
with the highest Emax per DDI object.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\deqn{f = 1 + \frac{(E_{max}) * C^n}{(EC_{50}^n + C^n)}}
}
\details{
Fit a 3-parameter hill function to the mRNA induction data from each
donor and DDI object.
}
\examples{
indmod(induction_experiment(examplinib_in_vitro_ind, "examplinib"))
}
