% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_tools.R
\name{indmod}
\alias{indmod}
\title{Analyze in vitro CYP induction data}
\usage{
indmod(x, precipitant = "")
}
\arguments{
\item{x}{In vitro induction data as data frame with the columns:
\itemize{
\item DONOR The hepatocyte donor
\item SAMPLE "test" or "positive_control"
\item CONC The preipitant concentration in uM.
\item OBJECT The DDI object.
\item FOLD The mRNA or activity fold change.
\item REL The fold change relative to positive control.
\item SOURCE Source information
}}

\item{precipitant}{The precipitant as character.}
}
\value{
A list with the elements:
\itemize{
\item data The original data, model and model parameters for each fit.
\item fold_plot A ggplot object showing the original data and the fits.
\item ind_param The ec50, hill (n) and emax parameters by donor and DDI object.
\item inducer An inducer object representing the model parameters from the donor
with the highest Emax per DDI object.
}
}
\description{
Fit a 3-parameter hill function to the mRNA induction data from each
donor and DDI object.
}
\details{
\deqn{f = 1 + \frac{(E_{max}) * C^n}{(EC_{50}^n + C^n)}}
}
\examples{
indmod(examplinib_in_vitro_ind, precipitant = "examplinib")
}
