% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inducer.R
\docType{class}
\name{inducer-class}
\alias{inducer-class}
\title{DDI CYP inducer data object}
\value{
Inducer object
}
\description{
DDI CYP inducer data object
}
\section{Slots}{

\describe{
\item{\code{precipitant}}{The name of the DDI perpetrator as character.}

\item{\code{data}}{data.frame.}
}}

