% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inducer.R
\docType{class}
\name{inducer-class}
\alias{inducer-class}
\title{Title}
\value{
Inducer object
}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{object}}{character.}

\item{\code{data}}{data.frame.}
}}

