% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/induction.R
\name{induction_data}
\alias{induction_data}
\title{induction_data constructor alias}
\usage{
induction_data(data = NULL, precipitant = "")
}
\arguments{
\item{data}{A data frame with the columns object, ki and source.}

\item{precipitant}{The precipitant name as character.}
}
\value{
An inhibition_data object.
}
\description{
induction_data constructor alias
}
