% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_vitro_inducer.R
\name{induction_experiment}
\alias{induction_experiment}
\title{constructor for induction_experiment objects}
\usage{
induction_experiment(data = NULL, precipitant = "")
}
\arguments{
\item{data}{A tibble.}

\item{precipitant}{The precipitant as character.}
}
\value{
An induction_experiment object.
}
\description{
constructor for induction_experiment objects
}
