% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inhibition.R
\name{inhibition_data}
\alias{inhibition_data}
\title{inhibition_data constructor}
\usage{
inhibition_data(data = NULL, precipitant = "")
}
\arguments{
\item{data}{A data frame with the columns object, ki and source.}

\item{precipitant}{The precipitant name as character.}
}
\value{
An inhibition_data object.
}
\description{
inhibition_data constructor
}
