% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inhibitor.R
\docType{class}
\name{inhibitor-class}
\alias{inhibitor-class}
\title{Direct inhibitor class definition}
\value{
An inhibitor class object.
}
\description{
Direct inhibitor class definition
}
\section{Slots}{

\describe{
\item{\code{object}}{Name of perpetrator.}

\item{\code{data}}{Data frame with the fields 'object', 'ki' and 'source'.}
}}

