% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inhibitor.R
\docType{class}
\name{inhibitor-class}
\alias{inhibitor-class}
\title{Inhibitor class definition}
\value{
An inhibitor class object.
}
\description{
Inhibitor class definition
}
\section{Slots}{

\describe{
\item{\code{precipitant}}{Name of DDI precipitant.}

\item{\code{data}}{Data frame with the fields 'object', 'ki' and 'source'.}
}}

