% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{is_igut_solubility_limited}
\alias{is_igut_solubility_limited}
\title{Test if Igut of a perpetrator is limited by its solubility}
\usage{
is_igut_solubility_limited(obj)
}
\arguments{
\item{obj}{The perpetrator object.}
}
\value{
A Boolean value.
}
\description{
This function tests whether the solubility of a perpetrator compound is
limiting its intestinal concentration.
}
\details{
If the 'solubility' field is \code{Inf} (default), or the compound solubility is
larger than \eqn{I_{gut}}, the function returns \code{FALSE}. If the solubility
is lower than the theoretical \eqn{I_{gut}}, i.e., lower than the dose
dissolved in 250 ml, the function returns \code{TRUE}. Note that the solubility
is expected in mg/l.
}
\examples{
is_igut_solubility_limited(examplinib_parent)
}
\seealso{
\code{\link[=key_concentrations]{key_concentrations()}}
}
