% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{is_igut_solubility_limited}
\alias{is_igut_solubility_limited}
\title{Test if Igut of a perpetrator is limited by its solubility}
\usage{
is_igut_solubility_limited(obj)
}
\arguments{
\item{obj}{The perpetrator object.}
}
\value{
A boolean value.
}
\description{
If the solubility field is `Inf` (default), or the compound solubility is
larger than Igut, the function returns `FALSE`. If the solubility is lower
than the theoretical Igut, i.e., lower than the dose dissolved in 250 ml,
the function returns `TRUE`. Note that the solubility is expected in mg/l.
}
