% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Item-class}
\alias{Item-class}
\title{Item, This class contains the information of an Item in the Dictonary}
\description{
Item, This class contains the information of an Item in the Dictonary
}

