% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipitant.R
\name{key_conc_table}
\alias{key_conc_table}
\title{Key precipitant concentrations}
\usage{
key_conc_table(x, round = 3, qh = 1.616, qent = 18/60)
}
\arguments{
\item{x}{DDI precipitant object.}

\item{round}{Number of decimal places.}

\item{qh}{Hepatic blood flow in L/min.}

\item{qent}{Enteric blood flow in L/min.}
}
\value{
Markdown-formatted text.
}
\description{
Print a markdown-formatted table of the relevant precipitant
concentrations in \eqn{\mu M} and ng/ml for a DDI precipitant compound.
}
\details{
\subsection{Gut concentration}{

\deqn{I_{gut} = \frac{D} {250}}

If the above exceeds the aqueous solubility of the drug, \eqn{I_{gut}} is set
to its solubility.
}

\subsection{Unbound systemic concentration}{

\deqn{I_{max,ss,u}=I_{max,ss} * f_u}
}

\subsection{Unbound hepatic inlet concentration}{

For orally administered (parent) compounds, the hepatic inlet concentration
is the systemic concentration plus a portal term:

\deqn{portal\ term = D*\frac{F_a*F_g*k_a}{Q_h*R_B}*1000\ ng/ml}

where \eqn{D} is the administered dose in mg, \eqn{F_a} the fraction absorbed
after oral administration, \eqn{F_g} the fraction available after gut
metabolism, \eqn{k_a} the absorption rate, \eqn{Q_h} the hepatic blood flow
and \eqn{R_B} the blood-to-plasma ratio.

The relevant hepatic inlet concentration (\eqn{I_{max,inlet,u}}, also called
\eqn{I_h} in the mechanistic static modeling equations) concentration is the
sum of the maximal systemic plasma concentration and the portal contribution:

\deqn{I_{max,inlet,u}=(I_{max,ss} + portal\ term) * f_u}
}

\subsection{Enteric (villous) concentration}{

For orally administered (parent) compounds, the villous concentration in the
gut (\eqn{I_{enteric}}, also called \eqn{I_g} in the mechanistic static
modeling equations) is calculated as:

\deqn{I_{enteric,u} = D * \frac{F_a*k_a}{Q_{ent}} *1000\ ng/ml}

where \eqn{F_a} is the fraction absorbed after oral administration, \eqn{k_a}
the absorption rate, \eqn{Q_{ent}} the enteric villous blood flow and
\eqn{R_B} the blood-to-plasma distribution ratio of the compound.

Note that as per the FDA guideline (refer to FDA, 2020, Fig. 7, and
Rostami-Hodjegan and Tucker, 2004) the blood-to-plasma ratio and the plasma
binding of the drug are not applicable.
}
}
\examples{
key_conc_table(examplinib)

}
