% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{key_concentrations}
\alias{key_concentrations}
\title{Key perpetrator concentrations}
\usage{
key_concentrations(obj, qh = 1.616, qent = 18/60, molar = TRUE)
}
\arguments{
\item{obj}{A perpetrator object.}

\item{qh}{Hepatic blood flow in l/min, defaults to 1.616 l/min.}

\item{qent}{Enteric blood flow in l/min, defaults to 0.3 l/min = 18 l/h.}

\item{molar}{Boolean value to select output in molar concentrations.}
}
\value{
Key perpetrator concentrations as a named vector.
}
\description{
Key perpetrator concentrations
}
