% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyp.R
\name{kinetic_cyp_induction_risk.list}
\alias{kinetic_cyp_induction_risk.list}
\title{Basic kinetic CYP induction risk}
\usage{
\method{kinetic_cyp_induction_risk}{list}(perp, cyp_ind)
}
\arguments{
\item{perp}{A list of perpetrator objects.}

\item{cyp_ind}{The CYP induction data as data frame.}
}
\value{
A list of data frames.
}
\description{
Basic kinetic CYP induction risk
}
\examples{
kinetic_cyp_induction_risk(examplinib_compounds, examplinib_cyp_induction_data)
}
