% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_assessment.R
\name{kinetic_cyp_induction_risk}
\alias{kinetic_cyp_induction_risk}
\title{Kinetic assessment of CYP induction risk}
\usage{
kinetic_cyp_induction_risk(perp, cyp_ind, d = 1)
}
\arguments{
\item{perp}{Precipitant object}

\item{cyp_ind}{induction_data object.}

\item{d}{Scaling factor, defaults to 1.}
}
\value{
DDI risk object
}
\description{
Kinetic assessment of CYP induction risk
}
\examples{
kinetic_cyp_induction_risk(examplinib, examplinib_cyp_induction)

}
