% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_assessment.R
\name{kinetic_cyp_induction_risk}
\alias{kinetic_cyp_induction_risk}
\title{Title}
\usage{
kinetic_cyp_induction_risk(perp, cyp_ind, d = 1)
}
\arguments{
\item{perp}{Perpetrator object}

\item{cyp_ind}{Induction object}

\item{d}{Numeric}
}
\value{
DDI risk object
}
\description{
Title
}
