% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyp.R
\name{kinetic_cyp_induction_risk}
\alias{kinetic_cyp_induction_risk}
\title{Generic function for basic kinetic CYP induction risk}
\usage{
kinetic_cyp_induction_risk(perp, cyp_ind)
}
\arguments{
\item{perp}{The perpetrator object or a list thereof.}

\item{cyp_ind}{The CYP induction data as data frame.}
}
\description{
Generic function for basic kinetic CYP induction risk
}
\examples{
kinetic_cyp_induction_risk(examplinib_parent, examplinib_cyp_induction_data)
}
