% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyp.R
\name{kinetic_cyp_induction_risk}
\alias{kinetic_cyp_induction_risk}
\title{Basic kinetic CYP induction risk}
\usage{
kinetic_cyp_induction_risk(perp, cyp_ind, d = 1)
}
\arguments{
\item{perp}{The perpetrator object or a list thereof.}

\item{cyp_ind}{The CYP induction data as data frame. The following fields
are expected:
\itemize{
\item 'name' The name of the perpetrator compound as character.
\item 'cyp' The CYP enzyme as (upper case) character.
\item 'emax' The \eqn{E_{max}}, i.e., the maximum induction effect determined in
vitro as numeric.
\item 'ec50' The \eqn{EC_{50}} in \eqn{\mu M} as numeric.
\item 'maxc' The maximal concentration in \eqn{\mu M} tested in the in vitro assay as
numeric.
\item 'source' Optional source information as character.
}}

\item{d}{Scaling factor, defaults to 1.}
}
\value{
A data frame.
}
\description{
Basic kinetic CYP induction risk
}
\details{
Basic kinetic modeling of the CYP induction risk considers:

\deqn{R = \frac {1}{1 + d * \frac {E_{max}*10*C_{max,ss,u}}{EC_{50,u} + 10 * C_{max,ss,u}}}}

\eqn{d} is a scaling factor with a standard value of 1. A different value can
be used if warranted by prior experience with the experimental conditions.

\eqn{R \le 0.8} suggest a relevant in vivo CYP induction potential.
}
\examples{
kinetic_cyp_induction_risk(examplinib_parent, examplinib_cyp_induction_data)
}
