% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyp.R
\name{kinetic_cyp_induction_risk}
\alias{kinetic_cyp_induction_risk}
\title{Basic kinetc CYP induction risk}
\usage{
kinetic_cyp_induction_risk(perp, cyp_ind)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{cyp_ind}{The CYP induction data as data frame.}
}
\value{
A data frame.
}
\description{
Basic kinetc CYP induction risk
}
