% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{load_cyp_inhibitor_data}
\alias{load_cyp_inhibitor_data}
\title{Load CYP inhibitor data}
\usage{
load_cyp_inhibitor_data(filename)
}
\arguments{
\item{filename}{The filename as string.}
}
\value{
The data as data frame.
}
\description{
Load CYP inhibitor data
}
