% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{load_perpetrators}
\alias{load_perpetrators}
\title{Load perpetrator file}
\usage{
load_perpetrators(filename)
}
\arguments{
\item{filename}{The full path to the compound file}
}
\value{
A list of perpetrator objects
}
\description{
Load perpetrator file
}
\details{
The file is expected to have the following content:
\itemize{
\item Comments start with '#' and go until the end of the line
\item lines have the format: \strong{drug}, \strong{parameter}, \strong{value}, \strong{source}
\item fields within a line are comma-separated
\item fields are plain text without enclosing ""
\item the source field is only for reference and usually refers to a clinical
or non-clinical study report. The source field may remain empty. Note
that the line then ends with a comma.
}

At least the following parameters are expected in the compound file\tabular{lll}{
   parameter \tab values \tab detail \cr
   type \tab 'parent' or 'metabolite' \tab the type of substance \cr
   mw \tab numeric \tab molar weight in g/mol \cr
   dose \tab numeric \tab clinical dose in mg \cr
   imaxss \tab numeric \tab clinical steady-state Cmax in ng/ml \cr
   fu \tab numeric \tab unbound fraction in plasma \cr
   fumic. \tab numeric \tab unbound fraction in microsomes \cr
   rb \tab numeric \tab blood to plasma ratio \cr
   fa \tab numeric \tab fraction absorbed \cr
   fg \tab numeric \tab fraction escaping gut metabolism \cr
   ka \tab numeric \tab absorption rate constant in 1/min \cr
}


For metabolites, the fields dose, fa, fg and ka are not relevant and should
be filled with \code{NA}.

Note that multiple compounds, e.g., the parent and metabolites may be
included in the perpetrator file. A typical compound file could look e.g.,
like this:

\if{html}{\out{<div class="sourceCode">}}\preformatted{# name, param, value, source
# parent

examplinib,  type,     parent,
examplinib,  mw,       492.6,
examplinib,  dose,     450,       clinical dose
examplinib,  imaxss,   3530,      study 001
examplinib,  fu,       0.023,     study 002
examplinib,  fumic,    1,         default
examplinib,  rb,       1,         study 003
examplinib,  fa,       0.81,      study 003
examplinib,  fg,       1,         default
examplinib,  ka,       0.00267,   unknown

# metabolite

M1,  type,   metabolite,
M1,  mw,     506.56,
M1,  dose,   NA,
M1,  imaxss, 1038,      study 001
M1,  fu,     0.012,     study 002
M1,  fumic,  1,         default
M1,  rb,     1,         study 002
M1,  fa,     NA,
M1,  fg,     NA,
M1,  ka,     NA,
}\if{html}{\out{</div>}}
}
\seealso{
\code{\link[=read_string_perpetrators]{read_string_perpetrators()}}
}
