% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_assessment.R
\name{mech_stat_cyp_risk}
\alias{mech_stat_cyp_risk}
\title{Title}
\usage{
mech_stat_cyp_risk(
  perp,
  cyp_inh,
  cyp_ind,
  cyp_tdi = NULL,
  d = 1,
  include_induction = TRUE,
  substr = cyp_reference_substrates,
  cyp_kdeg = cyp_turnover,
  qh = 1.616,
  qent = 18/60
)
}
\arguments{
\item{perp}{Perpetrator object}

\item{cyp_inh}{Inhibitor object}

\item{cyp_ind}{Inducer object}

\item{cyp_tdi}{Inhibitor object}

\item{d}{Numeric}

\item{include_induction}{Logical}

\item{substr}{Data frame}

\item{cyp_kdeg}{Data frame}
}
\value{
DDI risk object
}
\description{
Title
}
