% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyp.R
\name{mech_stat_cyp_risk}
\alias{mech_stat_cyp_risk}
\title{CYP perpetration risk as per mechanistic-static modeling}
\usage{
mech_stat_cyp_risk(
  perp,
  cyp_inh,
  cyp_ind,
  include_induction = T,
  substr = cyp_reference_substrates
)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{cyp_inh}{CYP inhibiton data as data frame.}

\item{cyp_ind}{CYP induction data as data frame.}

\item{include_induction}{Boolean value to define whether induction should be
included in the calculation (C-terms as per the FDA guideline)}

\item{substr}{The CYP reference substrates to be used as data frame.}
}
\value{
A data frame.
}
\description{
CYP perpetration risk as per mechanistic-static modeling
}
\details{
In this approach, AUC ratios for probe substrates are calculated based on
their known intestinal and hepatic metabolism. Both direct (competitive) and
time-dependent inhibition are considered.

The below formula given by the FDA guideline (refer to FDA 2020, Fig. 7) also
includes intestinal and hepatic enzyme induction terms (\eqn{C_g} and
\eqn{C_h}, respectively). At the same time, the guideline states that both
inhibition and induction should be considered separately.

\deqn{AUCR = \frac{1}{A_g*B_g*C_g* \left(1-F_g \right)+F_g} * \frac{1}{A_h*B_h*C_h*f_m+\left(1-f_m\right)}}

The individual terms are:
\subsection{Reversible inhibition}{

\deqn{A_g = \frac{1}{1+\frac{I_g}{K_i}}}

\deqn{A_h = \frac{1}{1+\frac{I_h}{K_i}}}
}

\subsection{Time-dependent inhibition}{

\deqn{B_g = \frac{k_{deg,g}}{k_{deg,g} + \frac{I_g*k_{inact}}{I_g+K_I}}}

\deqn{B_h = \frac{k_{deg,h}}{k_{deg,h} + \frac{I_h*k_{inact}}{I_h+K_I}}}
}

\subsection{Induction}{

\deqn{C_g = 1 + \frac{d*E_{max}*I_g}{I_g+EC_{50}}}

\deqn{C_h = 1 + \frac{d*E_{max}*I_h}{I_h+EC_{50}}}

with the hepatic inlet concentration \eqn{I_h=I_{max,inlet,u}} and the
intestinal concentration \eqn{I_g=I_{enteric,u}}, see
\code{\link[=key_concentrations]{key_concentrations()}}.

\eqn{d} is an induction scaling factor (assumed to
be 1, but can be adjusted).
}
}
\examples{
mech_stat_cyp_risk(examplinib_parent, examplinib_cyp_inhibition_data, examplinib_cyp_induction_data)

}
