% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statements.R
\name{mech_stat_cyp_risk_summary.perpetrator}
\alias{mech_stat_cyp_risk_summary.perpetrator}
\title{Summary statement for mechanistic-static modeling of CYP modulation}
\usage{
\method{mech_stat_cyp_risk_summary}{perpetrator}(
  perp,
  cyp_inh,
  cyp_ind = NULL,
  cyp_tdi = NULL,
  d = 1,
  include_induction = TRUE,
  substr = cyp_reference_substrates,
  cyp_kdeg = cyp_turnover
)
}
\value{
Markdown-formatted text output.
}
\description{
Summary statement for mechanistic-static modeling of CYP modulation
}
