% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statements.R
\name{mech_stat_cyp_risk_summary}
\alias{mech_stat_cyp_risk_summary}
\title{Summary statement for mechanistic-static modeling of CYP modulation}
\usage{
mech_stat_cyp_risk_summary(
  perp,
  cyp_inh,
  cyp_ind,
  cyp_tdi = NULL,
  d = 1,
  include_induction = TRUE,
  substr = cyp_reference_substrates,
  cyp_kdeg = cyp_turnover
)
}
\value{
Markdown-formatted text output.
}
\description{
Summary statement for mechanistic-static modeling of CYP modulation
}
\examples{
mech_stat_cyp_risk_summary(examplinib_parent,
  examplinib_cyp_inhibition_data, examplinib_cyp_induction_data)
mech_stat_cyp_risk_summary(examplinib_metabolite,
  examplinib_cyp_inhibition_data, examplinib_cyp_induction_data)
mech_stat_cyp_risk_summary(examplinib_compounds,
  examplinib_cyp_inhibition_data, examplinib_cyp_induction_data)
}
