% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyp.R
\name{mech_stat_cyp_risk_table}
\alias{mech_stat_cyp_risk_table}
\title{CYP perpetration risk table as per mechanistic-static modeling}
\usage{
mech_stat_cyp_risk_table(
  perp,
  cyp_inh,
  cyp_ind,
  substr = cyp_reference_substrates,
  na.rm = FALSE
)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{cyp_inh}{CYP inhibiton data as data frame.}

\item{cyp_ind}{CYP induction data as data frame.}

\item{substr}{The CYP reference substrates to be used as data frame.}

\item{na.rm}{Remove rows with lacking ki data (i.e., where ki == NA).}
}
\value{
A markdown-formatted table.
}
\description{
CYP perpetration risk table as per mechanistic-static modeling
}
