% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyp.R
\name{mech_stat_cyp_risk_table}
\alias{mech_stat_cyp_risk_table}
\title{CYP perpetration risk table as per mechanistic-static modeling}
\usage{
mech_stat_cyp_risk_table(
  perp,
  cyp_inh,
  cyp_ind,
  cyp_tdi = NULL,
  include_induction = TRUE,
  substr = cyp_reference_substrates,
  cyp_kdeg = cyp_turnover,
  na.rm = FALSE,
  show_dose = FALSE
)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{cyp_inh}{CYP inhibition data as data frame. The following fields are
expected:
\itemize{
\item 'name' The name of the perpetrator compound.
\item 'cyp' The CYP enzyme as (upper case) character.
\item 'ki' The \eqn{K_i} in \eqn{\mu M} as numeric.
\item 'source' Optional source information as character.
}}

\item{cyp_ind}{The CYP induction data as data frame. The following fields
are expected:
\itemize{
\item 'name' The name of the perpetrator compound as character.
\item 'cyp' The CYP enzyme as (upper case) character.
\item 'emax' The \eqn{E_{max}}, i.e., the maximum induction effect determined in
vitro as numeric.
\item 'ec50' The \eqn{EC_{50}} in \eqn{\mu M} as numeric.
\item 'maxc' The maximal concentration in \eqn{\mu M} tested in the in vitro assay as
numeric.
\item 'source' Optional source information as character.
}}

\item{cyp_tdi}{The CYP TDI data as data frame. The following fields are expected:
\itemize{
\item 'name' The perpetrator compound name as character.
\item 'cyp' The CYP enzyme as character.
\item 'ki' The \eqn{K_I} in \eqn{\mu M} as numeric.
\item 'kinact' The \eqn{k_{inact}} in 1/h as numeric.
\item 'source' Optional source information as character,
}}

\item{include_induction}{Switch to define whether induction effects should be
included in the calculation (C-terms as per the FDA guideline)}

\item{substr}{The CYP probe substrates to be used as data frame, defaults to
\link{cyp_reference_substrates}. The data frame is expected to have the following
fields:
\itemize{
\item 'cyp' The CYP enzyme as (upper case) character.
\item 'substrate' The substrate name as character.
\item 'fgut' The fraction of the drug escaping gut metabolism.
\item 'fm' The fraction of the drug that undergoes hepatic metabolism.
\item 'fmcyp' The fraction metabolized by the respective CYP enzyme.
}}

\item{cyp_kdeg}{The CYP turnover data as data frame. Defaults to the
built-in reference data, \link{cyp_turnover}.}

\item{na.rm}{Switch to define whether rows with lacking \eqn{K_I} or
\eqn{k_{deg}} data are removed from the output.}

\item{show_dose}{Show perpetrator dose in table title, defaults to \code{FALSE.}}
}
\value{
A markdown-formatted table.
}
\description{
CYP perpetration risk table as per mechanistic-static modeling
}
\examples{
mech_stat_cyp_risk_table(
  examplinib_parent,
  examplinib_cyp_inhibition_data,
  examplinib_cyp_induction_data)
mech_stat_cyp_risk_table(
  examplinib_compounds,
  examplinib_cyp_inhibition_data,
  examplinib_cyp_induction_data,
  examplinib_cyp_tdi_data)
mech_stat_cyp_risk_table(
  examplinib_compounds,
  examplinib_cyp_inhibition_data,
  examplinib_cyp_induction_data,
  examplinib_cyp_tdi_data, show_dose = TRUE)
}
\seealso{
\code{\link[=mech_stat_cyp_risk]{mech_stat_cyp_risk()}}

\link{cyp_reference_substrates}
}
