% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyp.R
\name{mech_stat_cyp_risk_table}
\alias{mech_stat_cyp_risk_table}
\title{CYP perpetration risk table as per mechanistic-static modeling}
\usage{
mech_stat_cyp_risk_table(
  perp,
  cyp_inh,
  cyp_ind,
  cyp_tdi = NULL,
  include_induction = T,
  substr = cyp_reference_substrates,
  kdeg = cyp_turnover,
  na.rm = FALSE
)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{cyp_inh}{CYP inhibiton data as data frame.}

\item{cyp_ind}{CYP induction data as data frame.}

\item{cyp_tdi}{CYP TDI data as data frame.}

\item{include_induction}{Boolean value to define whether induction should be
included in the calculation (C-terms as per the FDA guideline)}

\item{substr}{The CYP reference substrates to be used as data frame.}

\item{kdeg}{The CYP turnover data as data frame. Defaults to the
built-in reference data.}

\item{na.rm}{Remove rows with lacking ki data (i.e., where ki == NA).}
}
\value{
A markdown-formatted table.
}
\description{
CYP perpetration risk table as per mechanistic-static modeling
}
\examples{
mech_stat_cyp_risk_table(
  examplinib_parent,
  examplinib_cyp_inhibition_data,
  examplinib_cyp_induction_data)
mech_stat_cyp_risk_table(
  examplinib_parent,
  examplinib_cyp_inhibition_data,
  examplinib_cyp_induction_data,
  examplinib_cyp_tdi_data)
}
\seealso{
\code{\link[=mech_stat_cyp_risk]{mech_stat_cyp_risk()}}
}
