% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{names_string}
\alias{names_string}
\title{Perpetrator names as comma-separated string}
\usage{
names_string(perps)
}
\arguments{
\item{perps}{The perpetrator objects as a list.}
}
\value{
The output as string.
}
\description{
Perpetrator names as comma-separated string
}
