% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{names_string}
\alias{names_string}
\title{Perpetrator names as comma-separated string}
\usage{
names_string(perps)
}
\arguments{
\item{perps}{perpetrator objects as a list.}
}
\value{
The output as string.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
