% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{new_perpetrator}
\alias{new_perpetrator}
\title{Perpetrator object constructor}
\usage{
new_perpetrator(df)
}
\arguments{
\item{df}{A data frame to be converted into a perpetrator object.}
}
\value{
A perpetrator object.
}
\description{
The following columns are expected in the input:
* name
* type
* mw
* dose
* imaxss
* fu
* fumic
* rb
* fa
* fg
* ka
}
