% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{new_perpetrator}
\alias{new_perpetrator}
\title{Perpetrator object constructor}
\usage{
new_perpetrator(df)
}
\arguments{
\item{df}{A data frame to be converted into a perpetrator object. See
'Details' for the expected format and fields.}
}
\value{
A perpetrator object.
}
\description{
This function converts a data frame into a perpetrator object.
}
\details{
The input is a data frame with the columns 'param', 'value' and source'.

Rows with the following parameters ('param') are expected in the input:\tabular{lll}{
   PARAM \tab REQUIREMENT \tab DESCRIPTION \cr
   name \tab required \tab  \cr
   type \tab optional \tab either 'parent' (default) or 'metabolite' \cr
   mw \tab required \tab molar weight in g/mol \cr
   dose \tab required for parent compounds \tab clinical dose in mg \cr
   imaxss \tab required \tab total Cmax in ng/ml \cr
   fu \tab optional \tab fraction unbound, default is 1 \cr
   fumic \tab optional \tab microsomal unbound fraction, dafault is 1 \cr
   rb \tab optional \tab blood-to-plasma concentration ratio, default is 1 \cr
   fa \tab optional \tab fraction absorbed, default is 1 \cr
   fg \tab optional \tab fraction escaping gut metabolism, default is 1 \cr
   ka \tab optional \tab absorption rate constant, default is 0.1 /min \cr
   solubility \tab optional \tab solubility of the compound in mg/l, defaults to \code{Inf} \cr
}


The following example is an example for a valid input data frame:

\preformatted{
        param      value        source
1        name examplinib
2        type     parent
3          mw      492.6
4        dose        450 clinical dose
5      imaxss       3530     study 001
6          fu      0.023     study 002
7       fumic          1       default
8          rb          1     study 003
9          fa       0.81     study 003
10         fg          1       default
11         ka    0.00267       unknown
12 solubility        Inf       default
}
}
\seealso{
\code{\link[=read_perpetrators]{read_perpetrators()}}
}
