% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perpetrator.R
\name{perpetrator}
\alias{perpetrator}
\title{perpetrator class constructor}
\usage{
perpetrator(
  name,
  oral,
  mw,
  dose,
  imaxss,
  fu = 1,
  fumic = 1,
  rb = 1,
  fa = 1,
  fg = 1,
  ka = 0.1,
  solubility = Inf,
  source = character(0)
)
}
\arguments{
\item{name}{Character.}

\item{oral}{Oral administration, as logical.}

\item{mw}{Molecular weight in g/mol.}

\item{dose}{Administered dose in mg.}

\item{imaxss}{Cmax at steady state in ng/ml.}

\item{fu}{Fraction unbound.}

\item{fumic}{Fraction unbound in microsomes.}

\item{rb}{blood cell distribution.}

\item{fa}{Fraction absorbed.}

\item{fg}{Fraction escaping gut metabolism.}

\item{ka}{Absorption constant in 1/min.}

\item{solubility}{Aqueous solubility in mg/l.}

\item{source}{Source information for parameters as named character vector}
}
\value{
perpetrator object
}
\description{
perpetrator class constructor
}
\examples{
perpetrator(
"examplinib", TRUE, 492.6, 450, 3530, fu = 0.023, fa = 0.81, ka = .00267,
source = c(dose = "clinical dose", imaxss = "study 001", fu = "study 002")
)

}
