% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{perpetrator}
\alias{perpetrator}
\title{Make a perpetrator object from compound data}
\usage{
perpetrator(
  name,
  dose,
  imaxss,
  mw,
  oral = TRUE,
  fu = 1,
  fumic = 1,
  rb = 1,
  fa = 1,
  fg = 1,
  ka = 0.1,
  solubility = Inf,
  source = character(0)
)
}
\arguments{
\item{name}{The compound name as character.}

\item{dose}{The clinical dose in mg.}

\item{imaxss}{The (total) steady-state Cmax in ng/ml.}

\item{mw}{The molar weight in g/mol.}

\item{oral}{Oral administration as Boolean. Defaults to \code{TRUE}.}

\item{fu}{The fraction unbound as numeric. Defaults to 1.}

\item{fumic}{The fraction unbound in the microsomes. Defaults to 1.}

\item{rb}{The blood-to-plasma concentration ratio. Defaults to 1.}

\item{fa}{The fraction absorbed. Defaults to 1.}

\item{fg}{The fraction escaping gut metabolism. Defaults to 1.}

\item{ka}{The absorption rate constant in /ml. Defaults to 0.1 /ml.}

\item{solubility}{The aqueus solubility in mg/l. Defaults to Inf,}

\item{source}{Source information for the parameters as names character
vector.}
}
\value{
A perpetrator object.
}
\description{
This function creates a perpetrator object from key compound data.
}
\examples{
perpetrator("test", 100, 1000, 500, source = c(mw = "study 002"))
}
\seealso{
\code{\link[=read_perpetrators]{read_perpetrators()}}
}
