% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{print,ddi_risk-method}
\alias{print,ddi_risk-method}
\title{Print ddi_risk object}
\usage{
\S4method{print}{ddi_risk}(x)
}
\arguments{
\item{x}{A \code{ddi_risk} object.}
}
\value{
Markdown-formatted text.
}
\description{
Print ddi_risk object
}
