% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inhibitor.R
\name{print,inhibitor-method}
\alias{print,inhibitor-method}
\title{Title}
\usage{
\S4method{print}{inhibitor}(x)
}
\arguments{
\item{inhibitor}{Inhibitor object}
}
\value{
Markdown-formatted text
}
\description{
Title
}
