% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perpetrator.R
\name{print,perpetrator-method}
\alias{print,perpetrator-method}
\title{Property table for perpetrator object}
\usage{
\S4method{print}{perpetrator}(x)
}
\arguments{
\item{x}{A \code{perpetrator} object.}
}
\value{
Markdown-formatted text
}
\description{
Property table for perpetrator object
}
