% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inhibition.R
\name{print.inhibition_data}
\alias{print.inhibition_data}
\title{Print method for inhibition_data objects}
\usage{
print.inhibition_data(x, ...)
}
\arguments{
\item{x}{The inhibition}

\item{...}{Further arguments.}
}
\value{
Nothing.
}
\description{
Print method for inhibition_data objects
}
