% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{print.perpetrator}
\alias{print.perpetrator}
\title{Print implementation for perpetrator objects}
\usage{
\method{print}{perpetrator}(x, ...)
}
\arguments{
\item{x}{The perpetrator object.}

\item{...}{Further parameters.}
}
\description{
Print implementation for perpetrator objects
}
