% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perpetrator.R
\name{prop,compound-method}
\alias{prop,compound-method}
\title{Property table for perpetrator object}
\usage{
\S4method{prop}{compound}(x)
}
\arguments{
\item{compound}{}
}
\description{
Property table for perpetrator object
}
