% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{property_table.perpetrator}
\alias{property_table.perpetrator}
\title{Table of perpetrator drug properties}
\usage{
\method{property_table}{perpetrator}(obj)
}
\arguments{
\item{obj}{The perpetrator object.}
}
\value{
A markdown-formatted table.
}
\description{
This function generates a markdown-formatted table of the key properties of
a perpetrator object.
}
\examples{
property_table(examplinib_parent)

}
