% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{property_table}
\alias{property_table}
\title{Generic function to display compound properties}
\usage{
property_table(obj)
}
\arguments{
\item{obj}{The object (compund object or list thereof)}
}
\description{
Generic function to display compound properties
}
