% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{property_table}
\alias{property_table}
\title{Table of perpetrator properties}
\usage{
property_table(obj)
}
\arguments{
\item{obj}{The perpetrator object.}
}
\value{
Perpetrator properties as markdown-formatted table.
}
\description{
Table of perpetrator properties
}
