% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{read_cyp_inhibitor_data}
\alias{read_cyp_inhibitor_data}
\title{Read CYP inhibitor data}
\usage{
read_cyp_inhibitor_data(source)
}
\arguments{
\item{source}{The connection to read from.}
}
\value{
The data as data frame.
}
\description{
Read CYP inhibitor data
}
