% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{read_inducer_data}
\alias{read_inducer_data}
\title{#' Read CYP inhibitor data
#'
#' @param source The connection to read from.
#'
#' @return The data as data frame.
#' @export
read_cyp_inhibitor_data <- function(source) {
  return(read_dmpk_data(source))
}
Read csv-formatted CYP inducer data}
\usage{
read_inducer_data(source)
}
\arguments{
\item{source}{The connection to read from.}
}
\value{
The data as data frame.
}
\description{
#' Read CYP inhibitor data
#'
#' @param source The connection to read from.
#'
#' @return The data as data frame.
#' @export
read_cyp_inhibitor_data <- function(source) {
  return(read_dmpk_data(source))
}
Read csv-formatted CYP inducer data
}
