% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{read_inhibitor_data}
\alias{read_inhibitor_data}
\title{Read csv-formatted CYP inhibition data}
\usage{
read_inhibitor_data(source)
}
\arguments{
\item{source}{The connection to read from.}
}
\value{
The data as data frame.
}
\description{
This function loads CYP inhibition data from a csv file. The expected fields
are (in this order) the compound name, the CYP enzyme, the Ki and the source
information for the respective data. The latter field may remain empty.
}
\details{
Comment lines must start with '#'.

A valid source is, e.g.,
\preformatted{
# PARENT
# name,     CYP,     Ki,   source
examplinib, CYP1A2,  NA,
examplinib, CYP2B6,  NA,
examplinib, CYP2C8,  11,   study 001
examplinib, CYP2C9,  13.5, study 001
examplinib, CYP2C19, 15,   study 001
examplinib, CYP2D6,  NA,
examplinib, CYP3A4,  12.5, study 001
# METABOLITE
# name,     CYP,     Ki,   source
M1,         CYP2C9,  4.4,  study 002
}
}
