% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{read_inhibitor_data}
\alias{read_inhibitor_data}
\title{Read csv-formatted CYP inhibition data}
\usage{
read_inhibitor_data(source)
}
\arguments{
\item{source}{The connection to read from.}
}
\value{
A data frame.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated in favor of \code{\link[=read_cyp_inhibitor_data]{read_cyp_inhibitor_data()}},
\code{\link[=read_ugt_inhibitor_data]{read_ugt_inhibitor_data()}} or \code{\link[=read_transporter_inhibitor_data]{read_transporter_inhibitor_data()}}.
}
\details{
This function loads CYP inhibition data from a csv file. The expected fields
are (in this order) the compound name, the CYP enzyme, the Ki and the source
information for the respective data. The latter field may remain empty.

Comment lines must start with '#'.

A valid source is, e.g.,

\if{html}{\out{<div class="sourceCode">}}\preformatted{
# PARENT
# compound, CYP, ki, source
examplinib, CYP1A2,  NA,
examplinib, CYP2B6,  NA,
examplinib, CYP2C8,  11,   study 001
examplinib, CYP2C9,  13.5, study 001
examplinib, CYP2C19, 15,   study 001
examplinib, CYP2D6,  NA,
examplinib, CYP3A4,  12.5, study 001

# METABOLITE
M1,         CYP2C9,  4.4,  study 002
}\if{html}{\out{</div>}}
}
