% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{read_perpetrators}
\alias{read_perpetrators}
\title{Read perpetrator information from file or text string}
\usage{
read_perpetrators(source)
}
\arguments{
\item{source}{The connection to read from.}
}
\value{
A list of perpetrator objects.
}
\description{
The input source can either be a file name as string or a text connection.
This can be used to read the compound data from a string.
}
\details{
The file is expected to comply with the following:
\itemize{
\item Comments start with '#' and go until the end of the line
\item lines have the format: \strong{drug}, \strong{parameter}, \strong{value}, \strong{source}
\item fields within a line are comma-separated
\item fields are plain text without enclosing ""
\item the source field is only for reference and usually refers to a clinical
or non-clinical study report. The source field may remain empty. Note
that the line then ends with a comma.
}

For metabolites, the fields dose, fa, fg and ka are not relevant and should
be filled with \code{NA}. For metabolites, 'oral' must be 'FALSE'.

Note that multiple compounds, e.g., the parent and metabolites may be
included in the perpetrator file. A typical compound file could look e.g.,
like this:

\preformatted{
  # name, param, value, source
  # parent

  examplinib,  oral,     TRUE,
  examplinib,  mw,       492.6,
  examplinib,  dose,     450,       clinical dose
  examplinib,  imaxss,   3530,      study 001
  examplinib,  fu,       0.023,     study 002
  examplinib,  fumic,    1,         default
  examplinib,  rb,       1,         study 003
  examplinib,  fa,       0.81,      study 003
  examplinib,  fg,       1,         default
  examplinib,  ka,       0.00267,   unknown

  # metabolite

  M1,  oral,   FALSE,
  M1,  mw,     506.56,
  M1,  dose,   NA,
  M1,  imaxss, 1038,      study 001
  M1,  fu,     0.012,     study 002
  M1,  fumic,  1,         default
  M1,  rb,     1,         study 002
  M1,  fa,     NA,
  M1,  fg,     NA,
  M1,  ka,     NA,
  }
}
\examples{
read_perpetrators(textConnection(examplinib_compounds_string))
}
