% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{read_tdi_data}
\alias{read_tdi_data}
\title{Read csv-formatted CYP TDI data}
\usage{
read_tdi_data(source)
}
\arguments{
\item{source}{The connection to read from.}
}
\value{
A data frame.
}
\description{
Read csv-formatted CYP TDI data
}
