% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{read_transporter_inhibitor_data}
\alias{read_transporter_inhibitor_data}
\title{Read transporter inhibition data}
\usage{
read_transporter_inhibitor_data(source)
}
\arguments{
\item{source}{The file or text connection to read from.}
}
\value{
A data frame.
}
\description{
Read transporter inhibition data from a file or text connection.
}
\details{
The following, comma-separated fields are expected (in this order):
\itemize{
\item 'name' The perpetrator compound name
\item 'cyp' The UGT enzyme as (upper case) character.
\item 'ki' The \eqn{k_i}
\item 'source' Optional source information as character.
}

Lines starting with '#' are considered comments and are not evaluated.

The following is an example of a valid input:
\preformatted{
# PARENT
# name,     transporter, IC50, source
examplinib, Pgp,         0.41, study 005
examplinib, BCRP,        1.9,  study 005
examplinib, OCT1,        2.3,  study 006
examplinib, OATP1B1,     177,  study 006
examplinib, OATP1B3,     35,   study 006
examplinib, OAT1,        271,
examplinib, OAT3,        300,
examplinib, BSEP,        12.8,
examplinib, OCT2,        67,   study 006
examplinib, MATE1,       3.6,  study 006
examplinib, MATE2k,      1.1,  study 006
}
}
\examples{
read_transporter_inhibitor_data(textConnection(examplinib_transporter_inhibition_string))
}
