% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk}
\alias{risk}
\title{DDI risk object constructor}
\usage{
risk(table, precipitant, title = NULL)
}
\arguments{
\item{table}{The risk table.}

\item{precipitant}{The precipitant.}

\item{title}{The title for printing.}
}
\value{
A risk object.
}
\description{
DDI risk object constructor
}
