% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inducer.R
\name{show,inducer-method}
\alias{show,inducer-method}
\title{Show DDI inducer object}
\usage{
\S4method{show}{inducer}(object)
}
\arguments{
\item{inducer}{DDI inducer object}
}
\value{
Nothing.
}
\description{
Show DDI inducer object
}
