% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inhibitor.R
\name{show,inhibitor-method}
\alias{show,inhibitor-method}
\title{Show DDI inhibitor object.}
\usage{
\S4method{show}{inhibitor}(object)
}
\arguments{
\item{inhibitor}{Inhibitor object.}
}
\value{
Nothing.
}
\description{
Show DDI inhibitor object.
}
