% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perpetrator.R
\name{show,perpetrator-method}
\alias{show,perpetrator-method}
\title{Show DDI perpetrator objects.}
\usage{
\S4method{show}{perpetrator}(object)
}
\arguments{
\item{object}{A \code{perpetrator} object.}
}
\value{
Nothing.
}
\description{
Show DDI perpetrator objects.
}
