% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_assessment.R
\name{static_cyp_induction_risk}
\alias{static_cyp_induction_risk}
\title{Static risk assessment for CYP induction}
\usage{
static_cyp_induction_risk(perp, cyp_ind)
}
\arguments{
\item{perp}{Precipitant object.}

\item{cyp_ind}{Induction object.}
}
\value{
DDI risk object
}
\description{
Static risk assessment for CYP induction
}
\examples{
static_cyp_induction_risk(examplinib, examplinib_cyp_induction)
}
