% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{static_cyp_induction_risk}
\alias{static_cyp_induction_risk}
\title{Title}
\usage{
static_cyp_induction_risk(perp, cyp_ind)
}
\arguments{
\item{perp}{Perpetrator object}

\item{cyp_ind}{Induction object.}
}
\value{
DDI risk object
}
\description{
Title
}
