% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyp.R
\name{static_cyp_induction_risk}
\alias{static_cyp_induction_risk}
\title{Basic static CYP induction risk}
\usage{
static_cyp_induction_risk(perp, cyp_ind)
}
\arguments{
\item{perp}{The perpetrator object or a list thereof.}

\item{cyp_ind}{The CYP induction data as data frame. The following fields
are expected:
\itemize{
\item 'name' The name of the perpetrator compound as character.
\item 'cyp' The CYP enzyme as (upper case) character.
\item 'emax' The \eqn{E_{max}}, i.e., the maximum induction effect determined in
vitro as numeric.
\item 'ec50' The \eqn{EC_{50}} in µM as numeric.
\item 'maxc' The maximal concentration in µM tested in the in vitro assay as
numeric.
\item 'source' Optional source information as character.
}}
}
\value{
A data frame.
}
\description{
Basic static CYP induction risk
}
\details{
The basic (EMA) or fold-change (FDA) methods evaluate whether the maximal
fold-change in mRNA expression is > 2-fold at the expected unbound hepatic
concentration of the drug.

Regarding the relevant drug concentrations, the FDA guidance suggests
considering \eqn{30*I_{max,ss,u}} while the EMA guidance considers
\eqn{50*I_{max,ss,u}} for hepatic and \eqn{0.1*I_{gut}} for intestinal
CYP enzyme induction. It is expected that the concentrations in the
respective in vitro assays cover these concentrations.
}
\examples{
static_cyp_induction_risk(examplinib_parent, examplinib_cyp_induction_data)
}
