% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyp.R
\name{static_cyp_induction_risk}
\alias{static_cyp_induction_risk}
\title{Basic static CYP induction risk}
\usage{
static_cyp_induction_risk(perp, cyp_ind)
}
\arguments{
\item{perp}{The perpetrator object or a list thereof.}

\item{cyp_ind}{The CYP induction data as data frame.}
}
\value{
A data frame.
}
\description{
Basic static CYP induction risk
}
\details{
The basic (EMA) or fold-change (FDA) methods evaluate whether the maximal
fold-change in mRNA expression is > 2-fold at the expected unbound hepatic
concentration of the drug.

Regarding the relevant drug concentrations, the FDA guidance suggests
considering \eqn{30*I_{max,ss,u}} while the EMA guidance considers
\eqn{50*I_{max,ss,u}} for hepatic and \eqn{0.1*I_{gut}} for intestinal
CYP enzyme induction. It is expected that the concentrations in the
respective in vitro assays cover these concentrations.
}
\examples{
static_cyp_induction_risk(examplinib_parent, examplinib_cyp_induction_data)
}
