% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyp.R
\name{static_cyp_induction_risk}
\alias{static_cyp_induction_risk}
\title{Basic static CYP induction risk}
\usage{
static_cyp_induction_risk(perp, cyp_ind)
}
\arguments{
\item{perp}{The perpetrator object or a list thereof.}

\item{cyp_ind}{The CYP induction data as data frame. The following fields
are expected:
\itemize{
\item 'name' The name of the perpetrator compound as character.
\item 'cyp' The CYP enzyme as (upper case) character.
\item 'emax' The \eqn{E_{max}}, i.e., the maximum induction effect determined in
vitro as numeric.
\item 'ec50' The \eqn{EC_{50}} in \eqn{\mu M} as numeric.
\item 'maxc' The maximal concentration in \eqn{\mu M} tested in the in vitro assay as
numeric.
\item 'source' Optional source information as character.
}}
}
\value{
A data frame.
}
\description{
Basic static CYP induction risk
}
\details{
The basic fold-change method evaluates whether the maximal fold-change in
mRNA expression for the CYP enzyme, relative to the vehicle control, is >
2-fold at concentrations up to 50-fold over the unbound Cmax,ss - provided
that concentration-dependent increases in mRNA were observed in vitro.

If in the in vitro experiments concentrations of 50-fold the unbound Cmax,ss
were not tested, this is noted in the output.

Also note that the ICH guideline further specifies that the positive control
used in the in vitro studies should result in at least 6-fold mRNA increases,
otherwise an induction risk cannot be ruled out if the mRNA increase for the
CYP enzyme is > 20\% of the positive control. This, as well as the condition
of concentration-dependent increase in mRNA increase (see above), must be
confirmed manually by the user.

For details, refer to section 2.1.4.1 of the
\href{https://www.ema.europa.eu/en/documents/scientific-guideline/ich-m12-guideline-drug-interaction-studies-step-5_en.pdf}{ICH M1s2 guideline}
}
\examples{
static_cyp_induction_risk(examplinib_parent, examplinib_cyp_induction_data)
}
