% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyp.R
\name{static_cyp_induction_risk}
\alias{static_cyp_induction_risk}
\title{Generic function for basic static CYP induction risk}
\usage{
static_cyp_induction_risk(perp, cyp_ind)
}
\arguments{
\item{perp}{The perpetrator object or a list thereof.}

\item{cyp_ind}{The CYP induction data as data frame.}
}
\value{
A data frame.
}
\description{
Generic function for basic static CYP induction risk
}
\examples{
static_cyp_induction_risk(examplinib_parent, examplinib_cyp_induction_data)
}
