% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyp.R
\name{static_cyp_induction_risk_table}
\alias{static_cyp_induction_risk_table}
\title{Generic function: Table of the basic static CYP induction risk}
\usage{
static_cyp_induction_risk_table(perp, cyp_ind, na.rm = F)
}
\arguments{
\item{perp}{The perpetrator object or list thereof}

\item{cyp_ind}{The CYP induction data as data frame.}

\item{na.rm}{Remove rows with lacking ki data (i.e., where ki == NA).}
}
\description{
Generic function: Table of the basic static CYP induction risk
}
