% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyp.R
\name{static_cyp_induction_risk_table}
\alias{static_cyp_induction_risk_table}
\title{Table of the basic static CYP induction risk}
\usage{
static_cyp_induction_risk_table(perp, cyp_ind, na.rm = F, show_dose = FALSE)
}
\arguments{
\item{perp}{The perpetrator object or a list thereof.}

\item{cyp_ind}{The CYP induction data as data frame. The following fields
are expected:
\itemize{
\item 'name' The name of the perpetrator compound as character.
\item 'cyp' The CYP enzyme as (upper case) character.
\item 'emax' The \eqn{E_{max}}, i.e., the maximum induction effect determined in
vitro as numeric.
\item 'ec50' The \eqn{EC_{50}} in µM as numeric.
\item 'maxc' The maximal concentration in µM tested in the in vitro assay as
numeric.
\item 'source' Optional source information as character.
}}

\item{na.rm}{Switch to define whether rows with lacking \eqn{E_{max}} data
are removed from the output. Defaults to \code{FALSE}.}

\item{show_dose}{Show perpetrator dose in table title, defaults to \code{FALSE.}}
}
\value{
A markdown-formatted table.
}
\description{
Table of the basic static CYP induction risk
}
\examples{
static_cyp_induction_risk_table(examplinib_parent,
  examplinib_cyp_induction_data)
static_cyp_induction_risk_table(examplinib_compounds,
  examplinib_cyp_induction_data)
static_cyp_induction_risk_table(examplinib_compounds,
  examplinib_cyp_induction_data, show_dose = TRUE)
}
\seealso{
\code{\link[=static_cyp_induction_risk]{static_cyp_induction_risk()}}
}
