% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyp.R
\name{static_cyp_induction_risk_table}
\alias{static_cyp_induction_risk_table}
\title{Table of the basic static CYP induction risk}
\usage{
static_cyp_induction_risk_table(perp, cyp_ind, na.rm = F)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{cyp_ind}{The CYP induction data as data frame.}

\item{na.rm}{Remove rows with lacking ki data (i.e., where ki == NA).}
}
\value{
A markdown-formatted table.
}
\description{
Table of the basic static CYP induction risk
}
\examples{
static_cyp_induction_risk_table(examplinib_parent, examplinib_cyp_induction_data)
}
\seealso{
\code{\link[=static_cyp_induction_risk]{static_cyp_induction_risk()}}
}
