% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_assessment.R
\name{transporter_inh_risk}
\alias{transporter_inh_risk}
\title{Drug transporter inhibition risk}
\usage{
transporter_inh_risk(
  perp,
  transporter_inh,
  transporter_ref = transporter_reference_data,
  qh = 1.616
)
}
\arguments{
\item{perp}{Perpetrator object.}

\item{transporter_inh}{Inhibitor object.}

\item{transporter_ref}{Data frame.}
}
\value{
DDI risk object.
}
\description{
Drug transporter inhibition risk
}
\details{
The metric for the assessment of transporter interactions is
\eqn{R=[I]/IC_{50}}.

The relevant perpetrator concentrations \eqn{[I]} and regulatory thresholds
of concern are:\tabular{lll}{
   I \tab transporter \tab threshold \cr
   \eqn{I_{gut}} \tab P-gp and BRCR when drugs are orally administered \tab 10 \cr
   \eqn{C_{max,ss,u}} \tab P-gp and BRCR when drugs are administered parenterally or for drug metabolites \tab 0.02 \cr
   \eqn{I_{max,inlet,u}} \tab hepatic basolateral transporters OCT1, OATP1B1 and OATP1B3 \tab 0.1 \cr
   \eqn{C_{max,ss,u}} \tab renal basolateral transporters OAT1, OAT3 and OCT2 \tab 0.1 \cr
   \eqn{C_{max,ss,u}} \tab apical transporters MATE1 and MATE2-K \tab 0.02 \cr
}
}
