% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{transporter_inh_risk}
\alias{transporter_inh_risk}
\title{Title}
\usage{
transporter_inh_risk(
  perp,
  transporter_inh,
  transporter_ref = transporter_reference_data
)
}
\arguments{
\item{transporter_ref}{}
}
\description{
Title
}
