% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transporter.R
\name{transporter_inhibition_risk}
\alias{transporter_inhibition_risk}
\title{Drug transporter inhibition risk}
\usage{
transporter_inhibition_risk(
  perp,
  transporter_inh,
  transporter_ref = transporter_reference_data
)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{transporter_inh}{Transporter inhibition data as data frame. The
following fields are expected:
\itemize{
\item 'name' The perpetrator compound name
\item 'cyp' The UGT enzyme as (upper case) character.
\item 'ic50' The \eqn{IC_{50}} of the inhibition in μM.
\item 'source' Optional source information as character.
}}

\item{transporter_ref}{Transporter reference data, see
\link{transporter_reference_data} for details.}
}
\value{
A data frame.
}
\description{
Drug transporter inhibition risk
}
\details{
The metric for the assessment of transporter interactions is
\eqn{R=[I]/IC_{50}}.

The relevant perpetrator concentrations \eqn{[I]} are:
\itemize{
\item \eqn{I_{gut}} for P-gp and BRCR when drugs are orally administered, with a
threshold for R of 10
\item \eqn{C_{max,ss,u}} for P-gp and BRCR when drugs are administered
parenterally or for drug metabolites, with a threshold for R of 0.02
\item \eqn{I_{max,inlet,u}} for the hepatic basolateral transporters OCT1,
OATP1B1 and OATP1B3, with a threshold for R of 0.1
\item \eqn{C_{max,ss,u}} for the renal basolateral transporters OAT1, OAT3 and
OCT2, with a threshold for R of 0.1
\item \eqn{C_{max,ss,u}} for the apical transporters MATE1 and MATE2-K, with a
threshold for R of 0.02
}
}
\examples{
transporter_inhibition_risk(examplinib_parent,
  examplinib_transporter_inhibition_data)
}
\seealso{
\code{\link[=transporter_inhibition_risk_table]{transporter_inhibition_risk_table()}}

\code{\link[=read_transporter_inhibitor_data]{read_transporter_inhibitor_data()}}
}
