% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transporter.R
\name{transporter_inhibition_risk}
\alias{transporter_inhibition_risk}
\title{Drug transporter inhibition risk}
\usage{
transporter_inhibition_risk(
  perp,
  transporter_ic50,
  transporter_ref = transporter_reference_data
)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{transporter_ic50}{Transporter inhibition IC50 as data frame.}

\item{transporter_ref}{Transporter reference data.}
}
\value{
A data frame.
}
\description{
Drug transporter inhibition risk
}
