% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transporter.R
\name{transporter_inhibition_risk}
\alias{transporter_inhibition_risk}
\title{Drug transporter inhibition risk}
\usage{
transporter_inhibition_risk(
  perp,
  transporter_ic50,
  transporter_ref = transporter_reference_data
)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{transporter_ic50}{Transporter inhibition IC50 as data frame.}

\item{transporter_ref}{Transporter reference data.}
}
\value{
A data frame.
}
\description{
Drug transporter inhibition risk
}
\details{
The relevant metric for the assessment of transporter interactions is
\eqn{R=[I]/K_i}. In in vitro transporter inhibition studies, \eqn{IC_{50}}
values are experimentally determined. Since the transporter substrate
concentration is usually kept very low in relation to \eqn{K_m} to
minimize passive permeation, \eqn{K_i = IC_{50}} can be assumed.

The relevant perpetrator concentrations \eqn{[I]} are:
\itemize{
\item \eqn{I_{gut}} for intestinal P-gp and BRCR
\item \eqn{I_{max,inlet,u}} for the hepatic basolateral transporters OCT1,
OATP1B1 and OATP1B3
\item \eqn{I_{max,ss,u}} for the renal basolateral transporters OAT1, OAT3 and
OCT2, as well as the apical transporters outside the intestinal mucosa, i.e.,
hepatic P-gp and BCRP, and MATE1, MATE2-k.
}

Note that the FDA and EMA guidelines differ in their definitions for the
thresholds for assumed clinically relevant effects.
}
\examples{
transporter_inhibition_risk(examplinib_parent, examplinib_transporter_inhibition_data)

}
