% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transporter.R
\name{transporter_inhibition_risk}
\alias{transporter_inhibition_risk}
\title{Drug transporter inhibition risk}
\usage{
transporter_inhibition_risk(
  perp,
  transporter_inh,
  transporter_ref = transporter_reference_data
)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{transporter_inh}{Transporter inhibition data as data frame. The
following fields are expected:
\itemize{
\item 'name' The perpetrator compound name.
\item 'cyp' The UGT enzyme as (upper case) character.
\item 'ic50' The \eqn{IC_{50}} of the inhibition in μM.
\item 'source' Optional source information as character.
}}

\item{transporter_ref}{Transporter reference data, see
\link{transporter_reference_data} for details.}
}
\value{
A data frame.
}
\description{
Drug transporter inhibition risk
}
\details{
The metric for the assessment of transporter interactions is
\eqn{R=[I]/IC_{50}}.

The relevant perpetrator concentrations \eqn{[I]} and regulatory thresholds
of concern are:\tabular{lll}{
   I \tab transporter \tab threshold \cr
   \eqn{I_{gut}} \tab P-gp and BRCR when drugs are orally administered \tab 10 \cr
   \eqn{C_{max,ss,u}} \tab P-gp and BRCR when drugs are administered parenterally or for drug metabolites \tab 0.02 \cr
   \eqn{I_{max,inlet,u}} \tab hepatic basolateral transporters OCT1, OATP1B1 and OATP1B3 \tab 0.1 \cr
   \eqn{C_{max,ss,u}} \tab renal basolateral transporters OAT1, OAT3 and OCT2 \tab 0.1 \cr
   \eqn{C_{max,ss,u}} \tab apical transporters MATE1 and MATE2-K \tab 0.02 \cr
}
}
\examples{
transporter_inhibition_risk(examplinib_parent,
  examplinib_transporter_inhibition_data)
}
\seealso{
\code{\link[=transporter_inhibition_risk_table]{transporter_inhibition_risk_table()}}

\code{\link[=read_transporter_inhibitor_data]{read_transporter_inhibitor_data()}}
}
