% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statements.R
\name{transporter_inhibition_risk_summary.list}
\alias{transporter_inhibition_risk_summary.list}
\title{Title}
\usage{
\method{transporter_inhibition_risk_summary}{list}(perp, ...)
}
\arguments{
\item{perp}{A perpetrator object.}

\item{...}{Further parameters.}
}
\value{
Text output.
}
\description{
Title
}
