% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transporter.R
\name{transporter_inhibition_risk_table}
\alias{transporter_inhibition_risk_table}
\title{Table of drug transporter inhibition risks}
\usage{
transporter_inhibition_risk_table(
  perp,
  transporter_ic50,
  transporter_ref = transporter_reference_data,
  na.rm = F
)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{transporter_ic50}{Transporter inhibition IC50 as data frame.}

\item{transporter_ref}{Transporter reference data.}

\item{na.rm}{Boolean to indicate whether rows with lacking ki data should be
removed from the output (i.e., where ki == NA).}
}
\value{
A markdown-formatted table.
}
\description{
Table of drug transporter inhibition risks
}
\examples{
transporter_inhibition_risk_table(examplinib_parent, examplinib_transporter_inhibition_data)
}
\seealso{
\code{\link[=transporter_inhibition_risk]{transporter_inhibition_risk()}}
}
