% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transporter.R
\name{transporter_inhibition_risk_table}
\alias{transporter_inhibition_risk_table}
\title{Table of drug transporter inhibition risks}
\usage{
transporter_inhibition_risk_table(
  perp,
  transporter_inh,
  transporter_ref = transporter_reference_data,
  na.rm = F
)
}
\arguments{
\item{perp}{The perpetrator object.}

\item{transporter_inh}{Transporter inhibition data as data frame. See
\code{\link[=transporter_inhibition_risk]{transporter_inhibition_risk()}} for details.}

\item{transporter_ref}{Transporter reference data.}

\item{na.rm}{Switch to exlcude rows with lacking \eqn{IC_{50}} data from the
output.}
}
\value{
A markdown-formatted table.
}
\description{
Table of drug transporter inhibition risks
}
\examples{
transporter_inhibition_risk_table(examplinib_parent, examplinib_transporter_inhibition_data)
}
\seealso{
\code{\link[=transporter_inhibition_risk]{transporter_inhibition_risk()}}
}
