% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transporter_reference_data}
\alias{transporter_reference_data}
\title{Drug transporter reference data}
\format{
\itemize{
\item 'transporter' The name of the drug transporter protein.
\item 'threshold' The regulatory threshold for clinically relevant interactions.
\item 'i' The precipitant concentration metric applicable for the interaction.
}
}
\source{
Section 2 of the \href{https://www.ema.europa.eu/en/documents/scientific-guideline/ich-m12-guideline-drug-interaction-studies-step-5_en.pdf}{ICH M12 guideline}.
}
\usage{
transporter_reference_data
}
\description{
Drug transporter reference data
}
\details{
\preformatted{
transporter , threshold , i
Pgp_int     , 10        , igut
Pgp_sys     , 0.02      , imaxssu
BCRP_int    , 10        , igut
BCRP_sys    , 0.02      , imaxssu
OATP1B1     , 0.1       , imaxinletu
OATP1B3     , 0.1       , imaxinletu
OAT1        , 0.1       , imaxssu
OAT3        , 0.1       , imaxssu
BSEP        , 0.1       , imaxssu
OCT2        , 0.1       , imaxssu
MATE1       , 0.02      , imaxssu
MATE2k      , 0.02      , imaxssu
}
}
\seealso{
\code{\link[=key_conc_table]{key_conc_table()}}

\code{\link[=transporter_inhibition_risk]{transporter_inhibition_risk()}}
}
\keyword{datasets}
