% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transporter_reference_data}
\alias{transporter_reference_data}
\title{Drug transporter reference data}
\format{
## `transporter_reference_data`
}
\source{
FDA and EMA guidelines.
}
\usage{
transporter_reference_data
}
\description{
Drug transporter reference data
}
\keyword{datasets}
