% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transporter_reference_data}
\alias{transporter_reference_data}
\title{Drug transporter reference data}
\format{
An object of class \code{data.frame} with 13 rows and 5 columns.
}
\source{
FDA and EMA guidelines.
}
\usage{
transporter_reference_data
}
\description{
Drug transporter reference data
}
\keyword{datasets}
